==> 01_prep_ches.Rmd <==
---
title: "Prepare CHES Data"
output: html_notebook
---

# Load in Packages

```{r}
#install.packages("pacman")
#pacman::p_install_gh("systats/binoculaR")
pacman::p_load(tidyverse, magrittr, haven, ggthemes, sjPlot, sjmisc, sjstats, binoculaR, janitor, here, Amelia, mlbench, psych, ggrepel, cluster, factoextra, rvest)

range01 <- function(x){(x - min(x, na.rm = T)) / (max(x, na.rm = T) - min(x, na.rm = T))}

```

# Load in Data

```{r}
ches <- haven::read_dta(here("data", "1999_2014_CHES.dta"))

length(unique(ches$party_id))

```

## extracting party and country names 

```{r}
#devtools::install_github("expersso/pdftables")

# pdftable_api <- "opj9i5owyg40"
# 
# pdftables::convert_pdf("codebooks/party_tables2.pdf",
#                        "codebooks/party_tables.csv",
#                        api_key = pdftable_api)

# pdftables::convert_pdf("codebooks/country_names.pdf",
#                        "codebooks/country_names.csv",
#                        api_key = pdftable_api)

party_tables <- read_csv("codebooks/party_tables.csv")


colnames(party_tables) <- party_tables[1,]

party_tables %<>% 
  janitor::clean_names() %>% 
  select(party_id, party_name, party_name_english) %>% 
  filter(party_name != "Party Name") %>% 
  filter(party_name_english != "Party Name (English)") %>% 
  na.omit()

ches <- party_tables %>% 
  mutate(party_id = as.numeric(party_id)) %>% 
  left_join(ches, by = "party_id")

country_names <- read_csv("codebooks/country_names.csv") %>% janitor::clean_names()

colnames(country_names) <- country_names[1,]

country_names <- rbind(country_names[,1:3], 
                       country_names[,4:6]) 

colnames(country_names) <- c("id", "cntry_short", "cntry")

country_names %<>% 
  na.omit() %>% 
  select(cntry_short, cntry)

ches %<>% 
  mutate(cntry_short = str_to_upper(to_label(country))) %>% 
  left_join(country_names, by = "cntry_short")
```

# Correlations and PCAs

```{r}
ches %>% 
  select(antielite_salience, position, corrupt_salience) %>% 
  na.omit() %>% 
  cor()

ches %>% 
  select(sociallifestyle, civlib_laworder, galtan, multiculturalism, lrgen) %>% 
  na.omit() %>% 
  cor()

ches %>% 
  select(sociallifestyle, civlib_laworder, galtan, multiculturalism, lrgen) %>% 
  pca()

ches %>% 
  select(sociallifestyle, 
         civlib_laworder, 
         galtan, 
         multiculturalism, 
         lrgen) %>% 
  fa()

ches %>% 
  mutate(antieu = 1-range01(position)) %>% 
  select(antielite_salience, 
         antieu,
         corrupt_salience,
         sociallifestyle, 
         civlib_laworder, 
         galtan, 
         multiculturalism, 
         lrgen) %>% 
  fa(2, rotate = "varimax", fm = "ml")


```


# Recoding

```{r}

# Attempt 1
# anti-elite, anti eu + additive
# anti-elite, anti eu + factor

# Attempt 2
# anti-elite, anti eu, corruption + additive
# anti-elite, anti eu, corruption  + factor

# Attempt 3
# sociallifestyle, civlib_laworder, galtan + additive
# sociallifestyle, civlib_laworder, galtan + factor

# Attempt 4
# sociallifestyle, civlib_laworder, galtan multiculturalism, + additive
# sociallifestyle, civlib_laworder, galtan, multiculturalism + factor

# Attempt 5
# sociallifestyle, civlib_laworder, galtan, lrgen + additive
# sociallifestyle, civlib_laworder, galtan, lrgen + factor

# Attempt 6
# sociallifestyle, civlib_laworder, galtan, multiculturalism, lrgen + additive
# sociallifestyle, civlib_laworder, galtan, multiculturalism, lrgen + factor

ches %<>% 
  mutate(antieu = 1 - range01(position)) %>% 
  select(antielite_salience, corrupt_salience, 
         antieu, sociallifestyle, 
         civlib_laworder, galtan, country, 
         party, family, multiculturalism, 
         lrgen, lrecon, nationalism, year,
         party_name, party_name_english, cntry) %>% 
  mutate_at(vars(antielite_salience, 
                 corrupt_salience, 
                 antieu, 
                 sociallifestyle, 
                 civlib_laworder, 
                 galtan, 
                 multiculturalism, 
                 lrgen,
                 nationalism,
                 lrecon), range01) %>% 
  mutate_at(vars(antielite_salience, 
                 corrupt_salience, 
                 antieu, 
                 sociallifestyle, 
                 civlib_laworder, 
                 galtan, 
                 multiculturalism, 
                 lrgen,
                 nationalism,
                 lrecon), as.numeric) %>% 
  mutate(party_cntry = paste0(to_label(country), "_", party)) %>% 
  mutate(family = to_label(family)) %>%
  filter(year > 2013)


ches_add <- ches %>% 
  mutate(populism_e_eu_corr = range01(antielite_salience + 
                                        corrupt_salience + 
                                        antieu)*100) %>%
  mutate(populism_e_eu = range01(antielite_salience + 
                                   antieu)*100) %>% 
  mutate(liberalism_s_c_g = range01(sociallifestyle + 
                                      civlib_laworder + 
                                      galtan)*100) %>% 
  mutate(liberalism_s_c_g_m = range01(sociallifestyle + 
                                      civlib_laworder + 
                                      galtan + 
                                      multiculturalism)*100) %>% 
  mutate(liberalism_s_c_g_lr = range01(sociallifestyle + 
                                      civlib_laworder + 
                                      galtan + 
                                      lrgen)*100) %>% 
  mutate(liberalism_s_c_g_lr_m = range01(sociallifestyle + 
                                      civlib_laworder + 
                                      galtan + 
                                      lrgen +
                                      multiculturalism)*100) %>% 
  select(party_cntry, 
         party_name, 
         party_name_english,
         populism_e_eu_corr:liberalism_s_c_g_lr_m)

ches_add %>% 
  arrange(party_cntry)
# 
# ches %>% 
#   ggplot(aes(liberalism_s_c_g_lr_m, 
#              populism_e_eu_corr, 
#              color = family)) + 
#   geom_point() +
#   geom_text_repel(aes(liberalism_s_c_g_lr_m, 
#                       populism_e_eu_corr, 
#                       label = party_cntry)) +
#   guides(text = F) +
#   ggthemes::theme_hc() +
#   ggthemes::scale_color_hc()


# ches_pop <- ches %>% 
#   filter(populism >= 60)
# 
# length(unique(ches_pop$country))
# length(unique(ches$country))


# ches_elite <- ches %>% 
#   filter(antielite_salience >= 7.5)
# 
# length(unique(ches_elite$country))
# length(unique(ches$country))



```

## Factor Attempt

```{r}
#ff = fa_dat
fa_dat <- ches %>% 
  select(antielite_salience, antieu, 
         sociallifestyle, civlib_laworder, 
         galtan) %>% 
  mutate_all(as.numeric) 

# Check out a factor analysis
fa_pop <- fa_dat %>% 
  fa(2, rotate = "varimax", fm = "ml")

# save(fa_dat, file = "paper/data/fa_dat.Rdata")
# save(fa_pop, file = "paper/data/fa_pop.Rdata")

fa_dat_fin <- fa_dat %>% 
  fa(2, rotate = "varimax", fm = "ml") %>% 
  predict(fa_dat) %>% 
  as.data.frame() %>% 
  mutate(populism_e_eu_fac = range01(ML2)*100) %>% 
  mutate(liberalism_s_c_g_fac = range01(ML1)*100) %>% 
  select(-ML2, -ML1) %>% 
  cbind(ches) 

  
#   #fa(1, rotate = "cluster")
# 
# HS.model <- 'pop =~ 1*antielite_salience + position_r + corrupt_salience
#              lib =~ 1*sociallifestyle + civlib_laworder + galtan + multiculturalism
#              sociallifestyle ~~ galtan
#              antielite_salience ~~ corrupt_salience'
#  
#  fit <- lavaan::cfa(HS.model, data=ss, estimator ="MLM")
#  
#  lavaan::summary(fit, fit.measures=TRUE, std = T)
#  
#  lavaan::modificationindices(fit)

```

# Clustering

## function

```{r}
source("helper_functions.R")

plot_coords(ches_add, 
            liberalism_s_c_g_lr_m, 
            populism_e_eu)
  
ggsave(file = here("images", 
                   "diffs", 
                   "liberalism_s_c_g_lr_m.png"), 
       height = 10, width = 10)



plot_coords(fa_dat_fin, 
            liberalism_s_c_g_fac, 
            populism_e_eu_fac)
  
ggsave(file = here("images", 
                   "diffs", 
                   "populism_liberalism_fac2.png"), 
       height = 10, width = 10)
```

## by hand

```{r}
set.seed(2018)
ches_cluster <- ches_add %>% 
  select(party_cntry, liberalism_s_c_g, populism_e_eu) %>% 
  rename(prog_trad = liberalism_s_c_g,
         populism = populism_e_eu) %>% 
  #na.omit() %>% 
  as.data.frame()

row_names <- ches_cluster$party_cntry

ches_cluster %<>% 
  select(-party_cntry) %>% 
  mutate_all(scale) 

rownames(ches_cluster) <- row_names

k3 <- kmeans(ches_cluster, centers = 4, nstart = 25)


gap_stat <- clusGap(ches_cluster, FUN = kmeans, nstart = 25,
                    K.max = 10, B = 50)
# Print the result
print(gap_stat, method = "firstmax")
fviz_gap_stat(gap_stat) +
  ggthemes::theme_hc() +
  ggthemes::scale_color_gdocs()

ggsave(file = here("text", "images", "optim_cluster.png"), 
        height = 6, width = 10)


# ches %>%
#   select(party_cntry, prog_trad, populism) %>% 
#   na.omit() %>% 
#   mutate(cluster = k3$cluster,
#          state = row.names(ches_cluster)) %>%
#   ggplot(aes(prog_trad, populism, color = factor(cluster))) + 
#   geom_point() +
#   geom_text_repel(aes(prog_trad, populism, label = party_cntry), 
#                   show.legend = F) +
#   guides(text = F) +
#   ggthemes::theme_hc() +
#   ggthemes::scale_color_gdocs() +
#   xlab("<<Progressivism - Traditionalism>>") + 
#   ylab("<<Establishment - Populism>>") +
#   guides(color=guide_legend("Cluster")) +
#   ggtitle("Classification of European Parties") +
#   labs(subtitle = "Based on K-Nearest Neighbour Clustering",
#        caption = "Source: CHES Data 1999 - 2014") +
#   theme(axis.text = element_text(size = 15),
#         axis.title = element_text(size = 15),
#         legend.text = element_text(size = 15), 
#         legend.title = element_text(size = 15), 
#         title = element_text(size = 17))
#   
# ggsave(file = here("images", "diffs", "populism_e_eu_corr.png"), 
#        height = 10, width = 10)


```

# Save Cluster Data

```{r}
ches_final <- ches_add %>% 
  rename(prog_trad = liberalism_s_c_g,
         populism = populism_e_eu) %>%
  select(party_cntry, 
         party_name, 
         party_name_english,
         populism, 
         prog_trad) %>%
 # na.omit() %>%
  mutate(cluster = k3$cluster) %>%
  mutate(cluster = case_when(
    cluster == 4 ~ "Traditionalist Populism",
    cluster == 1 ~ "Progressive Populism",
    TRUE ~ "Establishment"
  ))

save(ches_final, file = "data/ches_final.Rdata")
#load("data/ches_pop3.Rdata")

populist_parties <- ches_final %>%
  filter(cluster != "Establishment") %>% 
  arrange(party_cntry)

save(populist_parties, file = "data/populist_parties.Rdata")
```

# making tables

```{r}
library(tidyr)
library(tidyverse)
ches_final %>% 
  select(-party_name, -party_name_english) %>% 
  left_join(ches, by = "party_cntry") %>% 
  # filter(!duplicated(populism, prog_trad))
  select(cntry, party_name, party_name_english, party_cntry, populism, prog_trad, cluster) %>% 
  mutate(party_name_english = str_remove(party_name_english, ";")) %>% 
  mutate(party_name_english = ifelse(str_detect(party_cntry, "be_SPA"), 
                                     "Socialistische Partij", 
                                     party_name_english)) %>%  
  # mutate(party_name = str_replace(party_name, "<U+0219>", "ș")) %>% 
  # mutate(party_name = case_when(
  #   str_detect(party_name, "formateur") ~ party_name_english,
  #   str_detect(party_name, "Parti Social Chr") ~ party_name_english,
  #   str_detect(party_name, "mocratique des Francophones") ~ party_name_english,    
  #   str_detect(party_name, "Partij van de Arbeid van Belgi") ~ party_name_english, 
  #   str_detect(party_name, "gelsen mod EU") ~ party_name_english,  
  #   str_detect(party_name, "90 Die Gr") ~ "Bündnis 90/Die Grünen", 
  #   str_detect(party_name, "Alternative f") ~ "Alternative für Deutschland",
  #   str_detect(party_name, "Panellinio Sosialistik") ~ "Panellinio Sosialistik Kinima",  
  #   str_detect(party_name, "a Dimokrat") ~ "Nea Demokratia",  
  #   str_detect(party_name, "Synaspism") ~ "SYRIZA ",  
  #   str_detect(party_name, "Kommounistik") ~ "Kommounistikó Kómma Elládas",  
  #   str_detect(party_name, "doxos Synagerm") ~ "Laikós Orthódoxos Synagermós",  
  #   str_detect(party_name, "Chrys") ~ "Golden Dawn",  
  #   str_detect(party_name, "Partido Socialista Obrero Espa") ~ "Partido Socialista Obrero Español",  
  #   str_detect(party_name, "ncia i Uni") ~ "Convergència i Unió",  
  #   str_detect(party_name, "n Canaria") ~ "Coalición Canaria",  
  #   str_detect(party_name, "n Progreso y Democracia") ~ "Unión Progreso y Democracia",  
  #   str_detect(party_name, "Partido de la Ciudadan") ~ "Ciudadanos - Partido de la Ciudadanía", 
  #   str_detect(party_name, "Parti Communiste Fran") ~ "Parti Communiste Français",  
  #   str_detect(party_name, "Rassemblement pour la R") ~ "Rassemblement pour la République",  
  #   str_detect(party_name, "Union pour la") ~ "Union pour la Démocratie Française", 
  #   str_detect(party_name, "Sinn F") ~ "Sinn Féin",     
  #   str_detect(party_name, "Fianna F") ~ "Fianna Fáil",  
  #   str_detect(party_name, "irti Lucht Oibre") ~ "Páirtí an Lucht Oibre",  
  #   str_detect(party_name, "isialach") ~ "Páirtí Sóisialach",  
  #   str_detect(party_name, "dtiroler Volkspartei") ~ "Südtiroler Volkspartei",  
  #   str_detect(party_name, "Sinistra Ecologia Libert") ~ "Sinistra Ecologia Libertà",  
  #   str_detect(party_name, "Fratelli d") ~ "Fratelli d'Italia",  
  #   str_detect(party_name, "Aoste") ~ "Vallée d'Aoste",  
  #   str_detect(party_name, "Coliga") ~ "Coligação Democrática Unitária",  
  #   str_detect(party_name, "Centro Democr") ~ "Centro Democrático e Social",  
  #   str_detect(party_name, "Partido Popular Democr") ~ "Partido Popular Democrático",  
  #   party_cntry == "aus_SPO" ~ "Sozialdemokratische Partei Österreichs",  
  #   str_detect(party_name, "sterreichische Volkspartei") ~ "Österreichische Volkspartei",  
  #   str_detect(party_name, "Freiheitliche Partei ") ~ "Freiheitliche Partei Österreichs",  
  #   party_cntry == "aus_GRUNE" ~ "Die Grünen",  
  #   str_detect(party_name, "ndnis Zukunft ") ~ "Bündnis Zukunft Österreich",  
  #   str_detect(party_name, "Team Stronach") ~ "Team Stronach für Österreich",  
  #   str_detect(party_name, " Liitto") ~ "Vihreä Liitto",  
  #   party_cntry == "fin_KD" ~ "Kristillisdemokraatit",     
  #   str_detect(party_name, "nsterpartiet") ~ "Vansterpartiet",  
  #   str_detect(party_name, "epartiet de Gr") ~ "Miljöpartiet de gröna",  
  #   str_detect(party_name, "enie za Prava i Svobodi") ~ "Dvizhenie za prava i svobodi",  
  #   str_detect(party_name, "enie Gergiovden") ~ "Dvijenie Gergiovden",  
  #   str_detect(party_name, "lne demokratick") ~ party_name_english,  
  #   str_detect(party_name, " Demokratick") ~ party_name_english,  
  #   str_detect(party_name, "strana Cech a Moravy") ~ "Komunistická strana Čech a Moravy",  
  #   str_detect(party_name, "Strana zelen") ~ "Strana zelených",  
  #   str_detect(party_name, "ANO 2011") ~ "ANO 2011",  
  #   str_detect(party_name, " demokracie") ~ "Úsvit - Národní koalice",  
  #   str_detect(party_name, "Strana svobodn") ~ "Svobodní",  
  #   str_detect(party_name, "Magyar Szocialista") ~ "Magyar Szocialista Párt", 
  #   str_detect(party_name, "Fidesz") ~ "Fidesz",  
  #   str_detect(party_name, "Lehet") ~ "Lehet Más a Politika",  
  #   str_detect(party_name, "tt 2014") ~ "Együtt",  
  #   str_detect(party_name, "Demokratikus Koal") ~ "Demokratikus Koalíció",  
  #   str_detect(party_name, "Nacionala apvieniba") ~ "Visu Latvijai!",  
  #   str_detect(party_name, "Saskanas Centrs") ~ party_name_english, 
  #   str_detect(party_name, "Latvijas Re") ~ party_name_english,  
  #   str_detect(party_name, "Lietuvos socialdemokratu") ~ party_name_english,  
  #   str_detect(party_name, "Tevynes sajunga") ~ party_name_english,  
  #   str_detect(party_name, "Lietuvos valstieciu") ~ party_name_english,  
  #   str_detect(party_name, "Lietuvos lenku") ~ party_name_english,  
  #   str_detect(party_name, "Lietuvos Respublikos Liberalu") ~ party_name_english,  
  #   str_detect(party_name, "Ruch Palikota") ~ "Ruch Palikota",  
  #   str_detect(party_name, "ionala pentru Progresul Rom") ~ party_name_english,  
  #   str_detect(party_name, "Uniunea Democrata Maghiara din Rom") ~ "Uniunea Democrată Maghiară din România",  
  #   str_detect(party_name, "Partidul Poporului") ~ "Partidul Poporului",  
  # 
  #   str_detect(party_name, " a krestansk") ~ "Slovenská demokratická a kresťanská únia", 
  #  party_cntry == "slo_SMK-MKP" ~ "Party of the Hungarian Community",
  #  party_cntry == "slo_KDH" ~ "Christian Democratic Movement",
  #  party_cntry == "slo_SNS" ~ "Slovak National Party",
  #  party_cntry == "slo_OLaNO" ~ "OLaNO",
  #  party_cntry == "slo_NOVA" ~ "New Majority",
  #  party_cntry == "cyp_DISY" ~ "Democratic Rally",
  #  party_cntry == "cyp_AKEL" ~ "Progressive Party of Working People",
  #  party_cntry == "cyp_DIKO" ~ "Democratic Party",
  #  party_cntry == "cz_KDU-CSL" ~ "Christian and Democratic Union",   
  #  str_detect(party_cntry, "UNPR") ~ "National Union for the Progress of Romania", 
  #  str_detect(party_cntry, "PMP") ~ party_name_english, 
  #   TRUE ~ party_name
  # )) %>% 
  mutate(party_name_english = iconv(party_name_english, "latin1", "ASCII", sub="")) %>% 
  # mutate(party_name = str_replace(party_name, "<U+FFFD>", "")) 
  mutate(populism = round(populism, 2)) %>% 
  mutate(prog_trad = round(prog_trad, 2)) %>% 
  arrange(cntry, cluster, desc(populism)) %>% 
  transmute(Country = cntry,
            `Party Code` = party_cntry,
            `Party Name` = party_name_english,
            `Populism Score` = populism,
            `Prog. - Trad. Score` = prog_trad,
            `Type` = cluster) -> partyscore_table

save(partyscore_table, file = "data/partyscore_table.Rdata")

library(xtable)
partyscore_xtable <- xtable::xtable(partyscore_table)
  
print(partyscore_xtable, booktabs = T)

```


# Crap


```{r}
ches %>% 
  ggplot(aes(liberalism, populism2, color = family)) + 
  geom_point() + 
  geom_hline(yintercept = 55, linetype = 4) +
  geom_text_repel(aes(liberalism, populism2, label = party_cntry)) +
  guides(text = F) +
  ggthemes::theme_hc() +
  ggthemes::scale_color_hc()

ggsave(file = here("images", "party_alignment3.png"), height = 10, width = 10)
```

## Function for getting website.

```{r}
getWebsite <- function(name)
{
    url = URLencode(paste0("https://www.google.com/search?q=",name))

    page <- read_html(url)

    results <- page %>%
      html_nodes("cite") %>% # Get all notes of type cite. You can change this to grab other node types.
      html_text()

    result <- results[1]

    return(as.character(result)) # Return results if you want to see them all.
}

fix_the_string <- function(string) {
  string %<>%
#  unique() %>%
  iconv(from = "Windows-1252", to = "UTF-8") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{4})>", "\\\\u$1") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{5})>", "\\\\U000$1") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{6})>", "\\\\U00$1") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{7})>", "\\\\U0$1") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{8})>", "\\\\U$1") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{1})>", "\\\\u000$1") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{2})>", "\\\\u00$1") %>% 
  stringi::stri_replace_all_regex("<U\\+([[:alnum:]]{3})>", "\\\\u0$1") %>% 
  stringi::stri_unescape_unicode() %>% 
  stringi::stri_enc_toutf8()
}

ches %<>%
  mutate(party_name = fix_the_string(party_name)) %>%
  mutate(party_name_english = fix_the_string(party_name_english)) 

# links <- ches$party_name %<>%
#   unique() %>%
#   paste(unique(ches$cntry), "english wikipedia") %>%
#   map_chr(getWebsite)
# 
# save(links, file = "data/party_links.Rdata")

load("data/party_links.Rdata")

merger <- ches$party_name %>%
  unique() %>%
  cbind(links) %>%
  data.frame()

colnames(merger) <- c("party_name", "links")

ches %<>%
  left_join(merger, by = "party_name")


table(ches$party_name)

# ches_pop2 <- ches %>% 
#   filter(populism2 >= 55) %>% 
#   mutate(lr_popul = ifelse(liberalism >= 50, 
#                            "Right-Wing Populism", 
#                            "Left-Wing Populism")) %>% 
#   mutate(links = links)
# 

#length(unique(ches_pop2$country))
length(unique(ches$country))

#dput(ches_pop2$party_cntry)

# ches_pop2 %<>% 
#   select(party_cntry, party_name, party_name_english, populism2, liberalism, 
#        electionyear, vote, seat, epvote, family, lr_popul, links)

# save(ches_pop2, file = "data/party_data.Rdata")

# 
# library(xml2) 
# library(httr) 
# library(rvest) 
# library(purrr) 
# library(dplyr)  
# 
# to_get <- seq(0, 150, 10) 
# pb <- progress_estimated(length(to_get))  
# 
# map_chr(to_get, function(i) {   
#   pb$tick()$print()   searchurl <- paste("http://pqasb.pqarchiver.com/djreprints/results.html?st=advanced&qrytxt=bankruptcy&sortby=chron&datetype=6&frommonth=01&fromday=01&fromyear=1908&tomonth=12&today=31&toyear=1908&by=&title=&at_hist=article&at_hist=editorial_article&at_hist=front_page&type=historic&start=", i, sep="")   htmlweb <- read_html(searchurl)   nodeweb <- html_node(htmlweb, "td > font.result_title > a")   
#   
# textweb <- html_text(nodeweb)   
# sys.sleep(sample(10, 1) * 0.1)   textweb }) -> titles 
# 
# print(trimws(titles))

table(ches$year)

```

## more Crap

```{r}
set.seed(123)
fviz_nbclust(ches_cluster, kmeans, method = "wss")
fviz_nbclust(ches_cluster, kmeans, method = "silhouette")


gap_stat <- clusGap(ches_cluster, FUN = kmeans, nstart = 25,
                    K.max = 10, B = 50)

fviz_gap_stat(gap_stat)
```

```{r}
ches_cluster <- ches %>% 
  select(party_cntry, antielite_salience, position, sociallifestyle, galtan, civlib_laworder) %>% 
  #na.omit() %>% 
  as.data.frame()

row_names <- ches_cluster$party_cntry

ches_cluster %<>% 
  select(-party_cntry) %>% 
  mutate_all(scale) 

rownames(ches_cluster) <- row_names

distance <- get_dist(ches_cluster)
png("images/distance_matrix.png", width = 1800, height = 1600)
fviz_dist(distance, 
                             gradient = list(low = "#00AFBB", 
                                             mid = "white",
                                             high = "#FC4E07"))
dev.off()

k3 <- kmeans(ches_cluster, centers = 4, nstart = 25)

fviz_cluster(k3, data = ches_cluster)

ches %>%
  select(party_cntry, liberalism, populism2) %>% 
  na.omit() %>% 
  mutate(cluster = k3$cluster,
         state = row.names(ches_cluster)) %>%
  ggplot(aes(liberalism, populism2, color = factor(cluster))) + 
  geom_point() +
  geom_text_repel(aes(liberalism, populism2, label = party_cntry)) +
  guides(text = F) +
  ggthemes::theme_hc() +
  ggthemes::scale_color_gdocs()

ggsave(file = here("images", "party_alignment6.png"), height = 10, width = 10)


```


```{r}
set.seed(123)
fviz_nbclust(ches_cluster, kmeans, method = "wss")
fviz_nbclust(ches_cluster, kmeans, method = "silhouette")


gap_stat <- clusGap(ches_cluster, FUN = kmeans, nstart = 25,
                    K.max = 10, B = 50)

fviz_gap_stat(gap_stat)
```




```{r}
ess <- read_spss("data/ess_round8.sav")


```

## Filter Data

```{r}
ches %>% 
  mutate(cntry = to_label(country))

# create a missing map
ches %>% 
  arrange(year, country) %>% 
  missmap(col=c("black", "grey"), legend=FALSE)

mod1 <- lm(antielite_salience ~ position, data = ches)

```
