==> 05_data_modeling.Rmd <==
---
title: "Data Modeling"
output: html_notebook
---

**TODO**
introduction
*appendix*
table of all parties and their score 
--> just make it fit (nooo change the UNICODES)
<!-- correlation table (geht net :/) -->


# Loading in Data and packages

```{r}
load("data/mlogit_ess.Rdata")

pacman::p_load(tidyverse, writexl, haven, sjPlot, sjmisc, texreg, psych, labelled, broom, magrittr, BaylorEdPsych, lmtest, datapasta, car, BBmisc, stargazer, caret, DescTools, rcompanion, tidyeval, mgcv, countrycode)

#pacman::p_install_gh("wilkelab/cowplot")
#devtools::install_github("gadenbuie/regexplain")

range01 <- function(x){(x - min(x, na.rm = T)) / (max(x, na.rm = T) - min(x, na.rm = T))}


# mlogit_ess %>% 
#   group_by(cntry, target) %>% 
#   tally() %>% #select(cntry) %>% unique()

```





# Sequentalism

## adds

```{r}

load("data/mlogit_ess_adds.Rdata")

mlogit_ess_adds %<>%
  na.omit()


library(nnet)
mlogit_ess_adds$populism2 <- relevel(mlogit_ess_adds$target, ref = "Establishment")

controls <- multinom(
  populism2 ~
    # economic deprivation
    #econ_unsat + econ_insec + unemployed + welfare +
    # cultural backlash
    #anti_imm + openness + conservation + selfenhance + selftrans +
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east + north + south +
    year_2016 + year_2014 + year_2012,
    #data
    data = mlogit_ess_adds, weights = pweight)

# controls_tidy <- step(controls) %>% 
#   tidy(exponentiate = T, conf.int = TRUE) %>% 
#   mutate(model = "control") %>% 
#   mutate(stars = get_p_values(p.value)) %>% 
#   mutate(mcfadden = round(PseudoR2(controls), 2)) 

  # cowplot::add_sub(gg_coefs, 
  #                  label = paste(expression(paste("McFadden's Pseudo R^2:")), 
  #                                           gg_coefs$mcfadden), 
  #                  x = 0, hjust = 0) %>%
  # cowplot::ggdraw()

fit1 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south + year_2016 + year_2014 + year_2012 +
    # economic deprivation
    econ_insec + unemployed + welfare,
    # cultural backlash
    #anti_imm + openness + conservation + selfenhance + selftrans + globalism + govsat +
    #data
    data = mlogit_ess_adds, weights = pweight)

# fit1_tidy <- step(fit1) %>% 
#   tidy(exponentiate = T, conf.int = TRUE) %>% 
#   mutate(model = "control") %>% 
#   mutate(stars = get_p_values(p.value)) %>% 
#   mutate(mcfadden = round(PseudoR2(fit1), 2)) 


fit2 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south +  year_2016 + year_2014 + year_2012+
    # economic deprivation
    #econ_unsat + econ_insec + unemployed + welfare +
    # cultural backlash
    anti_imm + openness + conservation + selfenhance + selftrans,
    #data
    data = mlogit_ess_adds, weights = pweight)


fit3 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south +   year_2016 + year_2014 + year_2012+
    # economic deprivation
    econ_insec + unemployed + welfare +
    # cultural backlash
    anti_imm + openness + conservation + selfenhance + selftrans,
    #data
    data = mlogit_ess_adds, weights = pweight)

fit4 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south +   year_2016 + year_2014 + year_2012+
    # economic deprivation
    econ_insec + unemployed + welfare +
    # cultural backlash
    anti_imm + openness + conservation + selfenhance + selftrans +
    # interactions
    econ_insec * anti_imm + 
    #econ_insec * conservation + 
    unemployed * selfenhance,
    #data
    data = mlogit_ess_adds, weights = pweight)

# PseudoR2(controls)
# PseudoR2(fit1)
# PseudoR2(fit2)
# PseudoR2(fit3)
# PseudoR2(fit4)
# #
# get_p_values <- function (pval) {
#   dplyr::case_when(
#     is.na(pval) ~ "", 
#     pval < 0.001 ~ "***", 
#     pval < 0.01 ~ "**", 
#     pval < 0.05 ~ "*", TRUE ~ "")
# }
# 
# 
# tidy(fit4) %>%
#   filter(term %nin% c("(Intercept)", "age", "educ", "sex",
#                      "lrscale", "ethnic", "religion", "rural",
#                      "globalism", "govsat", "east", "north", "south",
#                      "year_2016", "year_2014", "year_2012", "econ_insec",
#                      "unemployed", "welfare", "anti_imm", "openness",
#                      "conservation", "selfenhance", "selftrans")) %>%
#   mutate(stars = get_p_values(p.value))

# econ_insec:anti_imm only traditionalist populism
# econ_insec:conservation only traditionalist populism - weak
# unemployed:selftrans only progressive populism - weak

```



## one big one plot

### one prep

```{r}
get_p_values <- function (pval) {
  dplyr::case_when(
    is.na(pval) ~ "", 
    pval < 0.001 ~ "***", 
    pval < 0.01 ~ "**", 
    pval < 0.05 ~ "*", TRUE ~ "")
}

prep_dat <- function(model, modelname) {
  tidy_model <- tidy(model, exponentiate = T, conf.int = TRUE) %>% #step(model) %>% 
  mutate(model = modelname) %>% 
  mutate(stars = get_p_values(p.value)) 
  
tidy_model$mcfadden <- round(PseudoR2(model), 2)
  
  return(tidy_model)
}


anovies <- c("-", paste0(
round(c(anova(controls, fit1)$`LR stat.`[2],
anova(controls, fit2)$`LR stat.`[2],
anova(fit2, fit3)$`LR stat.`[2],
anova(fit3, fit4)$`LR stat.`[2]), 2),

c(anova(controls, fit1)$`Pr(Chi)`[2],
anova(controls, fit2)$`Pr(Chi)`[2],
anova(fit2, fit3)$`Pr(Chi)`[2],
anova(fit3, fit4)$`Pr(Chi)`[2]) %>% 
  get_p_values()
))



# # screenreg(test)
#  stargazer::stargazer(controls, fit1, fit2, fit3, type = "html", apply.coef = exp, p.auto=FALSE)

all_mods <- rbind( 
prep_dat(controls, "control"),
prep_dat(fit1, "model1"),
prep_dat(fit2, "model2"),
prep_dat(fit3, "model3"),
prep_dat(fit4, "model4")
) 

mcfaddies <- all_mods %>% 
  group_by(model, mcfadden) %>% 
  tally() %>%
  .$mcfadden

unique(all_mods$term)
```


### making the plot

```{r}



control_variables <- c("age", "educ", "sex", "lrscale", "ethnic", 
  "religion", "rural", "south", "east", "north", 
             "globalism", "govsat")
  
economic <- c("econ_insec", "unemployed", "welfare")

cultural <- c("anti_imm", "openness", 
             "conservation", "selfenhance", "selftrans")

all_mods %>% 
  filter(term != "south") %>% 
  filter(term != "east") %>% 
  filter(term != "north") %>% 
  filter(term != "year_2016") %>% 
  filter(term != "year_2014") %>% 
  filter(term != "year_2012") %>% 
  filter(term != "(Intercept)") %>% 
  mutate(variable_groups = case_when(
    term %in% control_variables ~ "Controls",
    term %in% economic ~ "Economic",
    term %in% cultural ~ "Cultural",
    TRUE ~ "Interactions"
  )) %>% 
  mutate(model = case_when(
    model == "control" ~ "Model 1 (Controls)",
    model == "model1" ~ "Model 2 (Economic)",
    model == "model2" ~ "Model 3 (Culture)",
    model == "model3" ~ "Model 4 (Economic + Culture)",
    model == "model4" ~ "Model 5 (Interactions)"
  )) %>%
  mutate(term = case_when(
    term == "anti_imm" ~ "Anti-Immigration Sentiment", 
    term == "openness" ~ "Openness", 
    term == "conservation" ~ "Conservation", 
    term == "selfenhance" ~ "Self-Enhancement", 
    term == "selftrans" ~ "Self-Transcendence", 
    term == "econ_insec" ~ "Economic Insecurity", 
    term == "unemployed" ~ "Unemployed (0/1)", 
    term == "welfare" ~ "Welfare (0/1)",
    term == "age" ~ "Age", 
    term == "educ" ~ "Education", 
    term == "sex" ~ "Female (0/1)", 
    term == "lrscale" ~ "Left-Right Scale", 
    term == "ethnic" ~ "Ethnic Minority (0/1)", 
    term == "religion" ~ "Religiosity", 
    term == "rural" ~ "Rural (0/1)",  
    term == "globalism" ~ "Trust in Global Governance", 
    term == "govsat" ~ "Government Satisfaction",
    term == "econ_insec:anti_imm" ~ "Econ. Insec. X Anti-Immigration",
#    term == "econ_insec:conservation" ~ "Economic Insecurity X Conservation",
    term == "unemployed:selfenhance" ~ "Unemployed X Self-Enhance"
  )) %>% 
  mutate(term = fct_relevel(term, levels = c("Unemployed X Self-Enhance",
                              #"Economic Insecurity X Conservation",
                              "Econ. Insec. X Anti-Immigration",
                              "Anti-Immigration Sentiment", "Openness", 
                              "Conservation", "Self-Enhancement", "Self-Transcendence", 
                              "Econ Insecurity", "Unemployed (0/1)", 
                              "Welfare (0/1)",
                              "Age", "Education", "Female (0/1)", "Left-Right Scale", 
                              "Ethnic Minority (0/1)", "Religiosity", "Rural (0/1)",
                              "Trust in Global Governance", "Government Satisfaction"))) %>% 
  mutate(`Variable Groups` = variable_groups) %>% 
  mutate(`Y-Level` = y.level) %>% 
  ggplot() + aes(x = term, y = estimate, 
                 ymin = conf.low, 
                 ymax = conf.high, 
                 color = `Y-Level`) + 
  geom_hline(yintercept = 1, 
             color = "gray25", 
             linetype = "dotted") + 
  geom_errorbar(position = position_dodge(0.5), width = 0) + 
  geom_point(aes(shape = `Variable Groups`), 
             position = position_dodge(width = 0.5)) + 
  scale_y_log10() +
  geom_vline(xintercept = seq(1.5, length(unique(all_mods$term)) - 0.5, 1), 
             lwd = 1, colour = "grey", alpha = 0.7) +
  geom_text(aes(x = term, 
                y = estimate + 0.01,
                label = paste(sprintf('%.2f', estimate, 2), stars)), 
                position = position_dodge(1.2), show.legend = F) +
  coord_flip() +
  guides(point = F) +
  scale_y_continuous(breaks = seq(-2,2, by = 0.3), limits = c(.5, 1.9)) +
  ggthemes::theme_hc() +
  ggthemes::scale_fill_fivethirtyeight() +
  cowplot::background_grid(major = 'y', minor = "none") + # add thin horizontal lines
  cowplot::panel_border() +# and a border around each panel
  xlab("") +
  ylab("Odds Ratios") + 
  # labs(caption = paste(expression("McFadden's Pseudo R^2:")), 
  #                                          gg_coefs$mcfadden) +
  facet_grid(~ model, scales = "fixed") +
  annotate("label", x = -1, y = 1.25, label = paste("McFadden's R2.:",
                                                  mcfaddies,
                                                  "\n N:", rep(nrow(mlogit_ess), 5),
                                                  "\n LRT:", anovies)) +
  # annotate("label", x = 7.5, y = 0.7, label = "Controls") +
  # annotate("label", x = 11.5, y = 0.7, label = "Economic") +
  # annotate("label", x = 18.5, y = 0.7, label = "Cultural") +
  # annotate("label", x = 19.5, y = 0.7, label = "Interactions") +
  # expand_limits(x = 21) + 
  geom_text(aes(x = -2, label = ""), show.legend = F) +
  scale_x_discrete(limits = c("Unemployed X Self-Enhance",
                              #"Economic Insecurity X Conservation",
                              "Econ. Insec. X Anti-Immigration",
                              "Interactions",
                              "Anti-Immigration Sentiment", "Openness", 
                              "Conservation", "Self-Enhancement", "Self-Transcendence", 
                              "Culture",
                              "Economic Insecurity", "Unemployed (0/1)", 
                              "Welfare (0/1)",
                              "Economics",
                              "Age", "Education", "Female (0/1)", "Left-Right Scale", 
                              "Ethnic Minority (0/1)", "Religiosity", "Rural (0/1)",
                              "Trust in Global Governance", "Government Satisfaction",
                              "Controls"), 
                   labels = c("Controls"=expression(displaystyle(bold(Controls))), 
                            "Economics"=expression(displaystyle(bold(Economics))),
                            "Culture"=expression(displaystyle(bold(Culture))), 
                            "Interactions"=expression(displaystyle(bold(Interactions))), 
                            parse=TRUE))

ggsave(filename = "images/onebigmotherfucker.png", width = 15, height = 10)
 
# multi1.rrr <- exp(coef(test))
# 
# stargazer::stargazer(test, type="text",coef=list(multi1.rrr), p.auto=FALSE)

# TODO Make a map (or two)
# TODO Make stuff citeable
# TODO Decide which interactions
# TODO Make probability plots

unique(all_mods$term)
```


# Citation Stuff

```{r}

cite_dat <- data.frame(anovies = anovies, model = unique(all_mods$model)) %>% 
  right_join(all_mods)

cite_dat %>% 
  filter(model == "control") %>% 
  mutate(id = 1:n()) %>% 
  select(y.level, term, estimate) %>% 
  spread("term", "estimate")

get_p_cite <- function (pval) {
  dplyr::case_when(is.na(pval) ~ "", 
                   pval < 0.001 ~ "< 0.001",
                   pval < 0.01 ~ "< 0.01", 
                   pval < 0.05 ~ "< 0.05", 
                   pval >= 0.05 ~ paste0("= ", sprintf('%.2f', pval)), 
                   TRUE ~ "")
}

cite_dat %<>% 
  mutate(estimate = sprintf('%.2f', estimate)) %>% 
  mutate(conf.low = sprintf('%.2f', conf.low)) %>% 
  mutate(conf.high = sprintf('%.2f', conf.high)) %>% 
  mutate(citation = paste0("OR = ", 
                          estimate, "; ",
                          "95% CI ", "[", conf.low, " - ", 
                          conf.high, "]; ", 
                          "p ", get_p_cite(p.value))) %>% 
  mutate(y.level = ifelse(y.level == "Progressive Populism", "prog", "trad"))

cite_complete <- function(mod, parameter, populism) {
  cite_dat %>%
    filter(model == mod) %>% 
    filter(term == parameter) %>% 
    filter(y.level == populism) %>% 
    select(citation) %>% .$citation
}

cite_OR <- function(mod, parameter, populism) {
  cite_dat %>%
    filter(model == mod) %>% 
    filter(term == parameter) %>% 
    filter(y.level == populism) %>% 
    select(estimate) %>% .$estimate 
}

cite_complete(mod = "control", 
              parameter = "age", 
              populism = "prog")

cite_OR(mod = "control", 
              parameter = "age", 
              populism = "prog")


save(cite_dat, file = "text/data/cite_dat.Rdata")
```


## Plot Predictions

### needed functions

```{r}
get_ames <- function(fit_eff) {
  df <- fit_eff$prob %>% 
  as.data.frame()

data.frame(diff(as.matrix(df))) %>% 
  #janitor::clean_names() %>% 
  summarise_all(mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%"))
}

prepper <- function(.data, wanted_var) {
  wanted_var <- enquo(wanted_var)
  
  data.frame(.data$model.matrix, 
           .data$prob, 
           .data$lower.prob, 
           .data$upper.prob) %>% 
  select(!!wanted_var, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, -!!wanted_var, 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>%
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
   mutate(scale = !!wanted_var) %>% 
   select(-!!wanted_var) 
}


from_to <- function(.data = mlogit_ess, variable) {
  variable <- enquo(variable)
  .data %>% 
    select(!!variable) %>% 
    table() %>% 
    names() %>% 
    as.vector() %>% 
    as.numeric()
} 
```


### controls

```{r}



  
as.numeric(as.vector(names(table(mlogit_ess$educ))))

fit_eff_age <- effects::effect("age", 
                           fit3)

fit_eff_educ <- effects::effect("educ", 
                           fit3)

fit_eff_sex <- effects::effect("sex", 
                           fit3)

# fit_eff_lrscale <- effects::effect("lrscale", 
#                            fit3, 
#                            xlevels = list(lrscale = mlogit_ess %>% from_to(lrscale)))
# 

fit_eff_ethnic <- effects::effect("ethnic", 
                           fit3)

fit_eff_religion <- effects::effect("religion", 
                           fit3)

# fit_eff_south <- effects::effect("south", 
#                            fit3)
# 
# fit_eff_east <- effects::effect("east", 
#                            fit3)
# 
# fit_eff_north <- effects::effect("north", 
#                            fit3)

fit_eff_globalism <- effects::effect("globalism", 
                           fit3)

fit_eff_govsat <- effects::effect("govsat", 
                           fit3)

rbind(
get_ames(fit_eff_age) %>% 
  mutate(term = "Age"),

get_ames(fit_eff_educ) %>% 
  mutate(term = "Education"),

get_ames(fit_eff_sex) %>% 
  mutate(term = "Female (0/1)"),

# get_ames(fit_eff_lrscale) %>% 
#   mutate(term = "Left-Right Scale"),

get_ames(fit_eff_ethnic) %>% 
  mutate(term = "Ethnic Minority (0/1)"),

get_ames(fit_eff_religion) %>% 
  mutate(term = "Religiosity"),

# get_ames(fit_eff_south) %>% 
#   mutate(term = "Southern Europe (0/1)"),
# 
# get_ames(fit_eff_east) %>% 
#   mutate(term = "Eastern Europe (0/1)"),
# 
# get_ames(fit_eff_north) %>% 
#   mutate(term = "Northern Europe (0/1)"),

get_ames(fit_eff_globalism) %>% 
  mutate(term = "Trust in Global Governance"),

get_ames(fit_eff_govsat) %>% 
  mutate(term = "Government Satisfaction")
) -> ame_dat_controls


rbind(
prepper(fit_eff_age, age) %>% 
  mutate(term = "Age"),

prepper(fit_eff_educ, educ) %>% 
  mutate(term = "Education"),

prepper(fit_eff_sex, sex) %>% 
  mutate(term = "Female (0/1)"),

# prepper(fit_eff_lrscale, lrscale) %>% 
#   mutate(term = "Left-Right Scale"),

prepper(fit_eff_ethnic, ethnic) %>% 
  mutate(term = "Ethnic Minority (0/1)"),

prepper(fit_eff_religion, religion) %>% 
  mutate(term = "Religiosity"),

# prepper(fit_eff_south, south) %>% 
#   mutate(term = "Southern Europe (0/1)"),
# 
# prepper(fit_eff_east, east) %>% 
#   mutate(term = "Eastern Europe (0/1)"),
# 
# prepper(fit_eff_north, north) %>% 
#   mutate(term = "Northern Europe (0/1)"),

prepper(fit_eff_globalism, globalism) %>% 
  mutate(term = "Trust in Global Governance"),

prepper(fit_eff_govsat, govsat) %>% 
  mutate(term = "Government Satisfaction")
) -> probs_dat_controls

cols <- c("#77AB43", "#FF2700", "#008FD5")

probs_dat_controls %>% 
  ggplot(aes(scale, value, colour = type)) +
    geom_smooth(method = "gam", 
               formula = y ~ s(x, k = 5), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    ggrepel::geom_label_repel(data = ame_dat_controls,
                              aes(x = Inf, y = Inf, label = V1, group = term),
              nudge_y = -10, show.legend = F, segment.alpha = 0
              #nudge_x = 15
              ) +
  #  geom_jitter(alpha = 0.01) +
  labs(y = "Marginal Probability") +
  # cowplot::background_grid(major = 'y', minor = "none") + # add thin horizontal lines
  guides(color = F) +
    ggthemes::theme_hc() +
  scale_color_manual(values = cols) +
    facet_grid(type~term, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 1),
                     labels = paste0(seq(0,100, by = 1),"%")) + 
  # cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) +
    # scale_x_continuous(breaks = seq(0,10, by = 1),
    #                  labels = seq(0,10, by = 1)) +
  xlab("Score") 


ggsave(filename = "text/images/control_probs.png", width = 18, height = 10)

```

#### extrawürschte

##### lrscale

```{r}

fit_eff_lrscale <- effects::effect("lrscale",
                           fit3,
                           xlevels = list(lrscale = mlogit_ess %>% from_to(lrscale)))



get_ames(fit_eff_lrscale) %>%
  mutate(term = "Left-Right Scale") %>% 
  mutate(`Y-Level` = type) -> ame_dat_lrscale


prepper(fit_eff_lrscale, lrscale) %>%
  mutate(term = "Left-Right Scale") -> probs_dat_lrscale


probs_dat_lrscale %>% 
  mutate(`Y-Level` = type) %>% 
  ggplot(aes(scale, value, colour = `Y-Level`)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 5), 
               size = 0.92, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper, group = `Y-Level`), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    ggrepel::geom_label_repel(data = ame_dat_lrscale,
                              aes(x = 9.2, y = 65, label = V1, group = term),
              nudge_y = -10, show.legend = F, direction = "y", segment.alpha = 0
              #nudge_x = 15
              ) +
  # geom_jitter(alpha = 0.70) +
  labs(y = "Marginal Probability") +
  # cowplot::background_grid(major = 'y', minor = "none") + # add thin horizontal lines
  #guides(color = T) +
    ggthemes::theme_hc() +
  scale_color_manual(values = cols) +
    #facet_grid(type~term, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 10),
                     labels = paste0(seq(0,100, by = 10),"%")) + 
  # theme(panel.spacing = unit(2, "lines")) +
    scale_x_continuous(breaks = seq(0,10, by = 1),
                     labels = seq(0,10, by = 1)) +
  xlab("Left-Right Scale") 


ggsave(filename = "text/images/lrscale_probs.png", width = 9, height = 6)
```

##### weirdo dummies

```{r}

  
as.numeric(as.vector(names(table(mlogit_ess$educ))))


fit_eff_south <- effects::effect("south",
                           fit3)

fit_eff_east <- effects::effect("east",
                           fit3)

fit_eff_north <- effects::effect("north",
                           fit3)

fit_eff_2012 <- effects::effect("year_2012", 
                           fit3) 


fit_eff_2014 <- effects::effect("year_2014", 
                           fit3)

fit_eff_2016 <- effects::effect("year_2016",  
                                fit3)


rbind(
get_ames(fit_eff_south) %>%
  mutate(term = "Southern Europe (0/1)"),

get_ames(fit_eff_east) %>%
  mutate(term = "Eastern Europe (0/1)"),

get_ames(fit_eff_north) %>%
  mutate(term = "Northern Europe (0/1)"),

get_ames(fit_eff_2012) %>% 
  mutate(term = "Year 2012 (0/1)"),

get_ames(fit_eff_2014) %>% 
  mutate(term = "Year 2014 (0/1)"),

get_ames(fit_eff_2016) %>% 
  mutate(term = "Year 2016 (0/1)")
) -> ame_dat_regyears


rbind(
prepper(fit_eff_south, south) %>%
  mutate(term = "Southern Europe (0/1)"),

prepper(fit_eff_east, east) %>%
  mutate(term = "Eastern Europe (0/1)"),

prepper(fit_eff_north, north) %>%
  mutate(term = "Northern Europe (0/1)"),

prepper(fit_eff_2012, year_2012) %>% 
  mutate(term = "Year 2012 (0/1)"),

prepper(fit_eff_2014, year_2014) %>% 
  mutate(term = "Year 2014 (0/1)"),

prepper(fit_eff_2016, year_2016) %>% 
  mutate(term = "Year 2016 (0/1)")
) -> probs_dat_regyears


probs_dat_regyears %>% 
  ggplot(aes(scale, value, colour = type)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 5), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    ggrepel::geom_label_repel(data = ame_dat_regyears,
                              aes(x = Inf, y = Inf, label = V1, group = term),
              nudge_y = -10, show.legend = F, segment.alpha = 0
              #nudge_x = 15
              ) +
  #  geom_jitter(alpha = 0.01) +
  labs(y = "Marginal Probability") +
  # cowplot::background_grid(major = 'y', minor = "none") + # add thin horizontal lines
  guides(color = F) +
    ggthemes::theme_hc() +
  scale_color_manual(values = cols) +
    facet_grid(type~term, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 2),
                     labels = paste0(seq(0,100, by = 2),"%")) + 
  #cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) +
    # scale_x_continuous(breaks = seq(0,10, by = 1),
    #                  labels = seq(0,10, by = 1)) +
  xlab("Score") 


ggsave(filename = "text/images/regyears_probs.png", width = 18, height = 10)
```


### cultural

#### all together

```{r}

fit.eff1 <- effects::effect("anti_imm", 
                           fit3, 
                           xlevels = list(anti_imm = c(1:10)))

fit.eff2 <- effects::effect("openness", 
                           fit3, 
                           xlevels = list(openness = c(1:10)))

fit.eff3 <- effects::effect("conservation", 
                           fit3, 
                           xlevels = list(conservation = c(1:10)))

fit.eff4 <- effects::effect("selfenhance", 
                           fit3, 
                           xlevels = list(selfenhance = c(1:10)))


fit.eff5 <- effects::effect("selftrans", 
                           fit3, 
                           xlevels = list(selftrans = c(1:10)))
rbind(
get_ames(fit.eff1) %>% 
  mutate(term = "Anti-Immigration Sentiment"),

get_ames(fit.eff2) %>% 
  mutate(term = "Openness"),

get_ames(fit.eff3) %>% 
  mutate(term = "Conservation"),

get_ames(fit.eff4) %>% 
  mutate(term = "Self-Enhancement"),

get_ames(fit.eff5) %>% 
  mutate(term = "Self-Transcendence")
) -> ame_dat


rbind(
prepper(fit.eff1, anti_imm) %>% 
  mutate(term = "Anti-Immigration Sentiment"),

prepper(fit.eff2, openness) %>% 
  mutate(term = "Openness"),

prepper(fit.eff3, conservation) %>% 
  mutate(term = "Conservation"),

prepper(fit.eff4, selfenhance) %>% 
  mutate(term = "Self-Enhancement"),

prepper(fit.eff5, selftrans) %>% 
  mutate(term = "Self-Transcendence")
) -> probs_dat

probs_dat %>% 
  ggplot(aes(scale, value, colour = type)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    ggrepel::geom_label_repel(data = ame_dat, 
                              aes(x = Inf, y = Inf, label = V1, group = term),
              nudge_y = -10, show.legend = F, segment.alpha = 0
              #nudge_x = 15
              ) +
  #  geom_jitter(alpha = 0.01) +
  labs(y = "Marginal Probability") +
  guides(color = F) +
    ggthemes::theme_hc() +
  scale_color_manual(values = cols) +
    facet_grid(type~term, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 1),
                     labels = paste0(seq(0,100, by = 1),"%")) +
    scale_x_continuous(breaks = seq(0,10, by = 1),
                     labels = seq(0,10, by = 1)) +
  #cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) +
  xlab("Score") 


ggsave(filename = "text/images/cultural_probs2.png", width = 18, height = 10)

```


### economic

```{r}


fit_eff_econ_insec <- effects::effect("econ_insec", 
                           fit3)

fit_eff_unemployed <- effects::effect("unemployed", 
                           fit3)

fit_eff_welfare <- effects::effect("welfare", 
                           fit3)

rbind(
get_ames(fit_eff_econ_insec) %>% 
  mutate(term = "Economic Insecurity"),

get_ames(fit_eff_unemployed) %>% 
  mutate(term = "Unemployed (0/1)"),

get_ames(fit_eff_welfare) %>% 
  mutate(term = "Welfare (0/1)")
) -> ame_dat_economic


rbind(
prepper(fit_eff_econ_insec, econ_insec) %>% 
  mutate(term = "Economic Insecurity"),

prepper(fit_eff_unemployed, unemployed) %>% 
  mutate(term = "Unemployed (0/1)"),

prepper(fit_eff_welfare, welfare) %>% 
  mutate(term = "Welfare (0/1)")
) -> probs_dat_economic


probs_dat_economic %>% 
  ggplot(aes(scale, value, colour = type)) +
    geom_smooth(method = "gam", 
               formula = y ~ s(x, k = 5), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    ggrepel::geom_label_repel(data = ame_dat_economic, 
                              aes(x = Inf, y = Inf, label = V1, group = term),
              nudge_y = -10, show.legend = F, segment.alpha = 0
              #nudge_x = 15
              ) +
  #  geom_jitter(alpha = 0.01) +
  labs(y = "Marginal Probability") +
  # cowplot::background_grid(major = 'y', minor = "none") + # add thin horizontal lines
  guides(color = F) +
    ggthemes::theme_hc() +
  scale_color_manual(values = cols) +
    facet_grid(type~term, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 1),
                     labels = paste0(seq(0,100, by = 1),"%")) + 
  # cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) +
    # scale_x_continuous(breaks = seq(0,10, by = 1),
    #                  labels = seq(0,10, by = 1)) +
  xlab("Score") 


ggsave(filename = "text/images/economic_probs.png", width = 18, height = 10)
```


## Plot Interactions

### Anti-Imm

```{r}
fit.eff <- effects::Effect(c("anti_imm", "econ_insec"), 
                           fit4, 
                           xlevels = list(anti_imm = c(0:10)), 
                                          econ_insec = c(1, 4))

get_ames_ints <- function(fit.eff) {
  df <- data.frame(fit.eff$prob,fit.eff$model.matrix) %>% 
    #filter(econ_insec %in% c(1,4)) %>%
    select(econ_insec, anti_imm, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism) %>% 
              filter(econ_insec %in% c(1,4))

  rbind(
data.frame(diff(as.matrix(df %>% filter(econ_insec == 1)))) %>% 
  # mutate(econ_insec = c(rep(1,11), rep(4,10))) %>% 
  # mutate(anti_imm = c(rep(0:10), rep(1:10))) %>% 
  # #janitor::clean_names() %>% 
  # group_by(econ_insec) %>% 
  summarise_at(vars(prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism), mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) %>% 
  mutate(econ_insec = 1),

data.frame(diff(as.matrix(df %>% filter(econ_insec == 4)))) %>% 
  summarise_at(vars(prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism), mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) %>% 
  mutate(econ_insec = 4)
)
}


get_ames_ints(fit.eff) %>% 
      mutate(`Economic Insecurity` = as.factor(econ_insec)) -> ame_int1


data.frame(fit.eff$model.matrix, 
           fit.eff$prob, 
           fit.eff$lower.prob, 
           fit.eff$upper.prob) %>% 
  select(anti_imm, econ_insec, # interaction effects
         prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, 
         -anti_imm, -econ_insec, # interaction effects 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>% 
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
  filter(econ_insec %in% c(1,4)) %>% 
  # mutate(Unemployed = as.factor(unemployed)) %>% 
  # mutate(unemployed = as.factor(case_when( 
  #   unemployed == max(unemployed) ~ max(unemployed),
  #   unemployed == min(unemployed) ~ min(unemployed),  
  #   TRUE ~ NA_real_
  # ))) %>% 
  #drop_na(unemployed) %>% 
  mutate(`Economic Insecurity` = as.factor(econ_insec)) %>% 
  ggplot(aes(anti_imm, value, group = `Economic Insecurity`, color = `Economic Insecurity`)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    scale_colour_manual("", values = c("blue", "red")) +
 #  geom_jitter(alpha = 0.01) +
    labs(y = "Marginal Probability") +
    ggrepel::geom_label_repel(data = ame_int1,
                              aes(x = Inf, y = Inf, 
                                  label = V1, group = type),
              nudge_y = -10, show.legend = F, segment.alpha = 0, direction = "y"
              #nudge_x = 15
              ) +
    ggthemes::theme_hc() +
    ggthemes::scale_color_fivethirtyeight() + 
    facet_wrap(~type, scales = "free") +
  scale_x_continuous(breaks = seq(0,10, by = 1)) +
    scale_y_continuous(breaks = seq(0,100, by = 1), 
                     labels = paste0(seq(0,100, by = 1),"%")) +
  xlab("Anti-Immigration Sentiment") +
  # ggtitle("Probabilities of Support for Establishment/Populist Parties") +
  # labs(subtitle = "Based on Multinomial Logistic Regression",
  #      caption = "Source: ESS Data Round 5 - 8") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 10),
        legend.text = element_text(size = 10), 
        legend.title = element_text(size = 10), 
        title = element_text(size = 12))


ggsave(file = here::here("images", "interaction_anti_imm.png"),
       height = 5, width = 10)
```

### interaction two

```{r}
fit.eff <- effects::Effect(c("selfenhance", "unemployed"), 
                           fit4, 
                           xlevels = list(selfenhance = c(0:10)), 
                                          unemployed = c(0, 1))

get_ames_ints <- function(fit.eff) {
  df <- data.frame(fit.eff$prob,fit.eff$model.matrix) %>% 
    #filter(econ_insec %in% c(1,4)) %>%
    select(selfenhance, unemployed, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism) %>% 
              filter(unemployed %in% c(0,1))

  rbind(
data.frame(diff(as.matrix(df %>% filter(unemployed == 0)))) %>% 
  # mutate(econ_insec = c(rep(1,11), rep(4,10))) %>% 
  # mutate(anti_imm = c(rep(0:10), rep(1:10))) %>% 
  # #janitor::clean_names() %>% 
  # group_by(econ_insec) %>% 
  summarise_at(vars(prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism), mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) %>% 
  mutate(unemployed = 0),

data.frame(diff(as.matrix(df %>% filter(unemployed == 1)))) %>% 
  summarise_at(vars(prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism), mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) %>% 
  mutate(unemployed = 1)
)
}


get_ames_ints(fit.eff) %>% 
      mutate(`Unemployed` = as.factor(unemployed)) -> ame_int2



data.frame(fit.eff$model.matrix, 
           fit.eff$prob, 
           fit.eff$lower.prob, 
           fit.eff$upper.prob) %>% 
  select(selfenhance, unemployed, # interaction effects
         prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, 
         -selfenhance, -unemployed, # interaction effects 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>% 
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
  filter(unemployed %in% c(0,1))  %>% 
      mutate(`Unemployed` = as.factor(unemployed)) %>% 
  ggplot(aes(selfenhance, value, group = `Unemployed` , color = `Unemployed`)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    scale_colour_manual("", values = c("blue", "red")) +
 #  geom_jitter(alpha = 0.01) +
    labs(y = "Marginal Probability") +
    ggrepel::geom_label_repel(data = ame_int2,
                              aes(x = Inf, y = Inf, 
                                  label = V1, group = type),
              nudge_y = -10, show.legend = F, segment.alpha = 0, direction = "y"
              #nudge_x = 15
              ) +
    ggthemes::theme_hc() +
    ggthemes::scale_color_fivethirtyeight() + 
    facet_wrap(~type, scales = "free") +
  scale_x_continuous(breaks = seq(0,10, by = 1)) +
    scale_y_continuous(breaks = seq(0,100, by = 1), 
                     labels = paste0(seq(0,100, by = 1),"%")) +
  xlab("Self-Enhancement Score") +
  # ggtitle("Probabilities of Support for Establishment/Populist Parties") +
  # labs(subtitle = "Based on Multinomial Logistic Regression",
  #      caption = "Source: ESS Data Round 5 - 8") +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(size = 10),
        legend.text = element_text(size = 10), 
        legend.title = element_text(size = 10), 
        title = element_text(size = 12))


ggsave(file = here::here("images", "interaction_unemployed.png"),
       height = 5, width = 10)

```

## Appropriateness

```{r}
head(pp <- fitted(fit3))
#head(pp <- predict(test, newdata = mlogit_ess, type = "class"))

conf.beta <- confint(object = fit3 , level = 0.95)
conf.beta # Results are in a 3D array

table(pp, mlogit_ess$populism2)

broom::tidy(test)

probs <- pp %>% 
  as_tibble() %>% 
  cbind(mlogit_ess) %>%
  as.data.frame()


 table(mlogit_ess$cntry)

probs %<>% 
  mutate(estab = ifelse(target == "Establishment", T, F)) %>% 
  mutate(lib = ifelse(target == "Progressive Populism", T, F)) %>% 
  mutate(illib = ifelse(target == "Traditionalist Populism", T, F)) 

table(probs$`Traditionalist Populism`>probs$`Progressive Populism` & probs$`Traditionalist Populism`>probs$Establishment, probs$illib)
table(probs$`Progressive Populism`>probs$`Traditionalist Populism` & probs$`Progressive Populism`>probs$Establishment, probs$lib)
table(probs$Establishment>probs$`Progressive Populism` & probs$Establishment>probs$`Traditionalist Populism`, probs$estab)
table(probs$populism2)

checkers <- probs %>% 
  filter(target != "Establishment")

table(checkers$`Traditionalist Populism`>checkers$`Progressive Populism`, checkers$illib)
table(checkers$`Progressive Populism`>checkers$`Traditionalist Populism`, checkers$lib)
table(probs$populism2)
```


## Predictions

```{r}


plot_probs <- function(probs = probs, variable = probs$anti_imm) {
  ss <- probs %>% 
  mutate(variable = as.numeric(variable)) %>% 
  select(variable, `Traditionalist Populism`, `Progressive Populism`, Establishment) %>% 
#  mutate(id = n()) %>% 
  gather(key = variable, value = value)  

names(ss) <- c("variable", "type", "value")

ss %>% 
  mutate(value = value * 100) %>% 
   ggplot(aes(variable, value, colour = type)) +
   geom_smooth(method = "gam", 
               formula = y ~ s(x, k = 1), 
               size = 1.2) +
 #  geom_ribbon(aes(ymin = LO, ymax = HI), fill = "gray70", alpha = 0.5) +
 #  geom_jitter(alpha = 0.01) +
  labs(x = str_glue("{names(ss)[1]}"), y = "Marginal Probability") +
  ggthemes::theme_hc() +
  ggthemes::scale_color_fivethirtyeight() + 
  facet_wrap(~type, scales = "free") +
  scale_y_continuous(breaks = seq(0,100, by = 5), 
                     labels = paste0(seq(0,100, by = 5),"%"))
}

plot_ints <- function(probs, variable = NULL, 
                      variable2 = NULL, type = "minmax", legend_title = "variable") {
#  if (type == "minmax") {

#}
  
ss <- probs %>% 
  mutate(variable = as.numeric(variable)) %>% 
  mutate(variable2 = as.numeric(variable2)) %>% 
  select(variable, variable2, `Traditionalist Populism`, 
         `Progressive Populism`, Establishment) %>% 
    mutate(variable2 = case_when(
    variable2 == max(variable2) ~ max(variable2),
    variable2 == min(variable2) ~ min(variable2),  
    TRUE ~ NA_real_
  )) %>% 
  drop_na(variable2) %>% 
#  mutate(id = n()) %>% 
  gather(key = "type", value = "value", -variable, -variable2) %>% 
  mutate(value = value * 100) 



# if (type == "sd") {
#   ss <- probs %>% 
#     mutate(variable2 = case_when(
#     variable2 == mean(variable2, na.rm = T) + 2*sd(variable2) max(variable2) ~ max(variable2),
#     variable2 == min(variable2) ~ min(variable2),  
#     TRUE ~ NA_real_
#   )) %>% 
#   drop_na(variable2) 
# }

ss %>% 
   ggplot(aes(variable, value, 
              colour = as.factor(variable2), group = variable2)) +
   geom_smooth(method = "gam", 
               formula = y ~ s(x, k = 1), 
               size = 1.2) +
 #  geom_jitter(alpha = 0.01) +
  labs(x = str_glue("{names(ss)[1]}"), y = "Marginal Probability") +
  ggthemes::theme_hc() +
#  scale_colour_hue("", l = 70, c = 150, h = c(90, 20)) + 
  scale_colour_manual(legend_title, values = c("blue", "red")) +
  facet_wrap(~type, scales = "free") +
  scale_y_continuous(breaks = seq(0,100, by = 5), 
                     labels = paste0(seq(0,100, by = 5),"%"))
}

plot_probs(probs, probs$religion)

ggsave(filename = "images/probabs.png", height = 5, width = 12)
# 
# variable <- mlogit_ess$econ_unsat
# variable2 <- mlogit_ess$anti_imm




plot_ints(probs, probs$anti_imm, probs$unemployed)

plot_ints(probs, probs$anti_imm, probs$unemployed, legend_title = "Unemployed over 3 Months") +
  xlab("Anti-Immigration Sentiment")

ggsave(filename = "images/interaction1.png", height = 5, width = 12)


plot_ints(probs, probs$anti_imm, probs$econ_insec, legend_title = "Economic Insecurity") +
  xlab("Anti-Immigration Sentiment")

ggsave(filename = "images/interaction2.png", height = 5, width = 12)


```

# Logistic Regression

```{r}
mlogit_ess %<>% 
  mutate(progressive_populism = ifelse(target == "Progressive Populism", 1, 0)) %>% 
  mutate(trad_populism = ifelse(target == "Traditionalist Populism", 1, 0))   

glmfit <- glm(progressive_populism ~ 
    # economic deprivation
    econ_unsat + econ_insec + unemployed +
    # cultural backlash
    anti_imm + authoritarianism + globalism +
    # control variables
    demsat + trust_gov + govsat + age + educ +  + sex ,
    #data
    data = mlogit_ess, weights = pweight, family = "binomial")

plot_model(glmfit)
screenreg(glmfit)
stargazer(glmfit, type = "text")


nagelkerke(glmfit)

plot_model(glmfit, type = "pred", terms = c("econ_unsat", "unemployed"))

predict(glmfit, mlogit_ess, type = "response") %>% 
  data.frame(predict_prog = .) %>% 
  transmute(predict_prog = predict_prog>=0.5) %>% 
  cbind(mlogit_ess) %>% 
  select(predict_prog, progressive_populism) %>% 
  table()

# Use your model to make predictions, in this example newdata = training set, but replace with your test set    
pdata <- predict(glmfit, newdata = mlogit_ess, type = "response")

# use caret and compute a confusion matrix
confusionMatrix(data = as.numeric(pdata>0.5), reference = mlogit_ess$progressive_populism)

glmfit <- glm(trad_populism ~ 
    # economic deprivation
    econ_unsat + econ_insec + unemployed +
    # cultural backlash
    anti_imm + authoritarianism + globalism +
    # control variables
    demsat + trust_gov + govsat + age + educ +  + sex ,
    #data
    data = mlogit_ess, weights = pweight, family = "binomial")

predict(glmfit, mlogit_ess, type = "response") %>% 
  data.frame(predict_trad = .) %>% 
  transmute(predict_trad = predict_trad>=0.5) %>% 
  cbind(mlogit_ess) %>% 
  select(predict_trad, trad_populism) %>% 
  table()

plot_model(glmfit)
screenreg(glmfit)

plot_model(glmfit, type = "pred", terms = c("econ_unsat"))
```




# Crap

## bayesian


```{r}
# controls <- multinom(
#   populism2 ~
#     # economic deprivation
#     #econ_unsat + econ_insec + unemployed + welfare +
#     # cultural backlash
#     #anti_imm + openness + conservation + selfenhance + selftrans +
#     # control variables
#     age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east + north + south +
#     year_2016 + year_2014 + year_2012,
#     #data
#     data = mlogit_ess_adds, weights = pweight)

library(brms)

br1 <- brm(populism2 ~ econ_unsat + econ_insec + unemployed + welfare + 
              anti_imm + age + educ + sex + lrscale + ethnic + religion + 
              rural + globalism + govsat + east + north + south +
              year_2016 + year_2014 + year_2012,
           data=mlogit_ess_adds, family="categorical",
           prior=c(set_prior ("normal (0, 8)")))

save(br1, file = "data/br1.Rdata")

Sys.getenv('PATH')
system('g++ -v')
system('where make')


load("data/br1.Rdata")

plot(br1) 


plot(marginal_effects(br1))
# 
# br1$ %>% 
#   as.data.frame()

# newdata <- mlogit_ess_adds %>% slice(1)
# predict(br1)

broom::tidy(br1)


devtools::install_github("mjskay/tidybayes")

library(tidybayes)

parameters(br1)

br1 %>%
  tidybayes::spread_samples(b_muTraditionalistPopulism_Intercept, b_muTraditionalistPopulism_anti_imm) %>%
  mean_qi(condition_mean = b_muTraditionalistPopulism_Intercept + b_muTraditionalistPopulism_anti_imm) %>%
  ggplot(aes(1, x = condition_mean, xmin = conf.low, xmax = conf.high)) +
  geom_halfeyeh() 

#mlogit_ess_adds %>%
 # modelr::data_grid("b_muTraditionalistPopulism_Intercept") %>%
  add_predicted_samples(mlogit_ess_adds, br1) -> pred_samples


```
## bayesian gerät

```{r}
mlogit_ess_adds %>% #select(target) %>% table 
  #filter(cntry == "Germany") %>% 
  sample_n(3000) -> mlogit_sub

library(brms)

br2 <- brm(populism2 ~ econ_unsat + econ_insec + unemployed + welfare + 
              anti_imm + age + educ + sex + lrscale + ethnic + religion + 
             anti_imm*unemployed +
              rural + globalism + govsat + #east + north + south +
              year_2016 + year_2014 + year_2012,
           data = mlogit_sub, family="categorical",
           prior = c(set_prior ("normal (0, 8)")))

save(br2, file = "data/br2.Rdata")
load("data/br2.Rdata")

library(tidybayes)

parameters(br2)

br2 %>%
  tidybayes::spread_samples(b_muTraditionalistPopulism_anti_imm,
                            b_muTraditionalistPopulism_govsat) %>%
  gather(key, value, -.chain, -.iteration) %>% 
  # mean_qi(condition_mean = b_muTraditionalistPopulism_Intercept +
  #           b_muTraditionalistPopulism_anti_imm) %>%
  ggplot(aes(y = value, x = key)) +
  geom_halfeyeh() 

add_predicted_samples(mlogit_sub, br2) -> pred_samples
add_fitted_samples(mlogit_sub, br2, n = 115) -> fit_samples

library(ggridges)

pred_samples %>%
  ggplot(aes(x = pred, y = anti_imm)) +
  geom_density_ridges()

fit_samples %>%
  ggplot(aes(x = anti_imm, y = estimate, color = category)) +
  geom_smooth(aes(y = estimate, group = paste(category, .iteration)), alpha = 0.25, 
              se = F, size = 0.000000000000025) +
  facet_wrap(~category, scales = "free") +
    ggthemes::theme_hc() +
  ggthemes::scale_colour_gdocs()

ggsave(filename = "images/bayesia_fun.png", width = 10)




sjPlot::plot_model(br2, 
                   terms = c("b_muTraditionalistPopulism_anti_imm",
                             "b_muTraditionalistPopulism_govsat"))
parameters(br2)
```

## bayesian2

```{r}
fit_samples %>%
  ungroup() %>% 
  select(govsat, globalism, lrscale, estimate, category, .iteration) %>% 
  gather("variable", "value", -estimate, -.iteration, -category) %>% 
  mutate(variable = 
           case_when(
             variable == "govsat" ~ "Government Satisfaction",
             variable == "globalism" ~ "Trust in Global Governance",
             variable == "lrscale" ~ "Left-Right Scale"
           )) %>% 
  mutate(category = 
           case_when(
             category == "1" ~ "Establishment",
             category == "2" ~ "Progressive Populism",
             category == "3" ~ "Traditionalist Populism"
           )) %>% 
  mutate(estimate = round(estimate * 100, 2)) %>% 
 ggplot(aes(x = value, y = estimate, color = category)) +
  geom_smooth(aes(y = estimate, group = paste(category, .iteration)), alpha = 0.25, 
              se = F, size = 0.000000000000025) +
  facet_grid(category~variable, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 5),
                     labels = paste0(seq(0,100, by = 5),"%")) + 
    ggthemes::theme_hc() +
    labs(y = "Probability", x = "Score") +
  ggthemes::scale_colour_gdocs("Type") + 
    guides(color = F) +
  # cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) 
    # scale_x_continuous(breaks = seq(0,10, by

ggsave(filename = "slides/images/bayesia_fun2.png", width = 12, height = 9)
```


## bayesian3

```{r}
#add_fitted_samples(mlogit_sub, br2, n = 100) -> fit_samples

fit_samples %>%
  ungroup() %>% 
  select(econ_insec, unemployed, welfare, estimate, category, .iteration) %>% 
  gather("variable", "value", -estimate, -.iteration, -category) %>% 
  mutate(variable =
           case_when(
             variable == "econ_insec" ~ "Economic Insecurity",
             variable == "unemployed" ~ "Unemployed (0/1)",
             variable == "welfare" ~ "Welfare (0/1)"
           )) %>%
  mutate(category = 
           case_when(
             category == "1" ~ "Establishment",
             category == "2" ~ "Progressive Populism",
             category == "3" ~ "Traditionalist Populism"
           )) %>% 
  mutate(estimate = round(estimate * 100, 2)) %>% 
 ggplot(aes(x = value, y = estimate, color = category)) +
  geom_smooth(aes(y = estimate, group = paste(category, .iteration)), alpha = 0.25, 
              se = F, size = 0.000000000000025, method = "gam") +
  facet_grid(category~variable, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 5),
                     labels = paste0(seq(0,100, by = 5),"%")) + 
    ggthemes::theme_hc() +
    labs(y = "Probability", x = "Score") +
  ggthemes::scale_colour_gdocs("Type") + 
    guides(color = F) +
  # cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) 
    # scale_x_continuous(breaks = seq(0,10, by

ggsave(filename = "slides/images/bayesia_fun3.png", width = 12, height = 9)
```

## bayesian4

```{r}

fit_samples %>%
  ungroup() %>% 
  select(anti_imm, religion, age, estimate, category, .iteration) %>% 
  gather("variable", "value", -estimate, -.iteration, -category) %>% 
  mutate(variable =
           case_when(
             variable == "anti_imm" ~ "Anti-Immigration Sentiment",
             variable == "religion" ~ "Religiosity",
             variable == "age" ~ "Age"
           )) %>%
  mutate(category = 
           case_when(
             category == "1" ~ "Establishment",
             category == "2" ~ "Progressive Populism",
             category == "3" ~ "Traditionalist Populism"
           )) %>% 
  mutate(estimate = round(estimate * 100, 2)) %>% 
 ggplot(aes(x = value, y = estimate, color = category)) +
  geom_smooth(aes(y = estimate, group = paste(category, .iteration)), alpha = 0.25, 
              se = F, size = 0.000000000000025, method = "loess") +
  facet_grid(category~variable, scales = "free") +
    scale_y_continuous(breaks = seq(0,100, by = 5),
                     labels = paste0(seq(0,100, by = 5),"%")) + 
    ggthemes::theme_hc() +
    labs(y = "Probability", x = "Score") +
  ggthemes::scale_colour_gdocs("Type") + 
    guides(color = F) +
  # cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) 
    # scale_x_continuous(breaks = seq(0,10, by

ggsave(filename = "slides/images/bayesia_fun4.png", width = 12, height = 9)
```



### example 

```{r}
set.seed(5)
n = 10
n_condition = 5
ABC =
  data_frame(
    condition = rep(c("A","B","C","D","E"), n),
    response = rnorm(n * 5, c(0,1,2,1,-1), 0.5)
  )

m = brm(response ~ (1|condition), data = ABC, control = list(adapt_delta = .99),
  prior = c(
    prior(normal(0, 1), class = Intercept),
    prior(student_t(3, 0, 1), class = sd),
    prior(student_t(3, 0, 1), class = sigma)
  ))

m %>%
  spread_samples(r_condition[condition,term]) %>%
  head(10)

m %>%
  spread_samples(b_Intercept, r_condition[condition,]) %>%
  mutate(condition_mean = b_Intercept + r_condition) %>%
  ggplot(aes(y = condition, x = condition_mean)) +
  geom_halfeyeh() 
```

## Interactions

```{r}
parameters(br2)

add_fitted_samples(mlogit_sub, br2, n = 100) -> fit_samples


fit_samples %>%
  ungroup() %>% 
  select(unemployed, anti_imm, estimate, category, .iteration) %>% 
  #gather("variable", "value", -estimate, -.iteration, -category, -unemployed) %>% 
  # mutate(variable = 
  #          case_when(
  #            variable == "unemployed" ~ "Unemployment (0/1)",
  #            variable == "anti_imm" ~ "Anti-Immigration Sentiment"
  # mutate(category = 
  #          case_when(
  #            category == "1" ~ "Establishment",
  #            category == "2" ~ "Progressive Populism",
  #            category == "3" ~ "Traditionalist Populism"
  #          )) %>% 
mutate(estimate = round(estimate * 100, 2)) %>% 
 ggplot(aes(x = anti_imm, y = estimate, color = unemployed)) +
  geom_smooth(aes(y = estimate, group = paste(unemployed, .iteration)), alpha = 0.25, 
              se = F, size = 0.000000000000025) +
  facet_grid(~category, scales = "free") +
    # scale_y_continuous(breaks = seq(0,100, by = 5),
    #                  labels = paste0(seq(0,100, by = 5),"%")) + 
    ggthemes::theme_hc() +
    labs(y = "Probability", x = "Score") +
  #ggthemes::scale_colour_gdocs("Type") + 
    guides(color = F) +
  # cowplot::panel_border() +# and a border around each panel
  theme(panel.spacing = unit(2, "lines"),
        strip.text = element_text(size = 11)) 
```


## Human Values

```{r}
# selected <- binoculaR::binoculaR(ess)
# 
# 
# fa_dat <- ess %>%
#   select(selected$var_codes)

#save(fa_dat, file = "data/fa_dat.Rdata")


# openness_dat <- fa_dat %>% 
#   select(ipcrtiv, impfree, impdiff, ipadvnt, ipgdtim, impfun) %>% 
#   psych::pca(1) 
#   #psych::alpha()
# 
# openness_dat %>% 
#   predict(., data = fa_dat)

# pacman::p_install_gh("yrosseel/lavaan")
# 
# fa_model <- 'openness =~ ipcrtiv + impfree + impdiff + ipadvnt + ipgdtim + impfun
#              ipgdtim ~~ impfun
#              ipcrtiv ~~ impfree
#              ipcrtiv ~~ impdiff
#              impfree ~~ ipadvnt'
# 
# 
# fit <- cfa(fa_model, data=fa_dat, 
#                    estimator= "MLM")
# 
# summary(fit, standardized=TRUE, 
#         fit.measures = TRUE, rsq = T)
# 
# openness <- lavaan::lavPredict(fit, newdata = fa_dat, fsm = T)
# 
# #Modifikationsindizes
# modheit <- modificationindices(fit, sort=T, minimum.value = 5)
# modheit
```

## crap 2

```{r}

library(rstan)

p <- predict(test)
table(p)
z <- summary(test)$coefficients/summary(test)$standard.errors
z

# 2-tailed z test
p <- (1 - pnorm(abs(z), 0, 1)) * 2
p

## extract the coefficients from the model and exponentiate
exp(coef(test))

head(pp <- fitted(test))

library(mlogit)
m <- mlogit(populism2 ~ imm_culture, data = mlogit_ess)
# compute a data.frame containing the mean value of the covariates in the sample
z <- with(mlogit_ess, data.frame(price = tapply(price, index(m)$alt, mean), 
	catch = tapply(catch, index(m)$alt, mean), 
econ_insec = mean(econ_insec)))
# compute the marginal effects (the second one is an elasticity
effects(m, covariate = "econ_insec", data = z)
effects(m, covariate = "price", type = "rr", data = z)
effects(m, covariate = "catch", type = "ar", data = z)

library(mnlogit)
#data(Fish, package = "mnlogit")
fm <- formula(populism2 ~ imm_culture )
fit <- mnlogit(fm, mlogit_ess, choiceVar="Establishment", ncores = 2)

### With improper prior it takes about 12 minutes, with about 40% CPU utilization and fans running,
### so you probably don't want to casually run the next line...

system.time (b1 <- brm (populism ~ imm_culture, 
                        data=mlogit_ess,
                        family="categorical", 
                        n.chains=3, 
                        n.iter=3000, 
                        n.warmup=600))
```



## Descriptives

```{r}

# pacman::p_install_gh("systats/binoculaR")
# 
# binoculaR::binoculaR(ess)

# ess %>% 
#   select(prtvede1, prtvede2) %>% na.omit() %>% mutate_all(sjmisc::to_label)

#var_label(ess$imueclt)

#ir <- data.frame (scale (iris[, -5]), Species=iris[, 5])


```


## Applying a different approach

```{r}

#install.packages("mlogit")    
# This is my favorite package to do multinomial logit in R

options(scipen=9)    # To fix notation
# library(MASS)
library(xtable)
library(mlogit)
library(reshape2)    # We will need this package at the very end


mnl.1 <- mlogit(populism2 ~ 1 |
    # economic deprivation
    econ_unsat + econ_insec + unemployed +
    # cultural backlash
    anti_imm + authoritarianism + globalism +
    # control variables
    demsat + trust_gov + govsat + age + educ + sex + lrscale + ethnic + religion + rural,
                data = mlogit_ess, shape = "wide", reflevel = "Establishment", weights = pweight)
summary(mnl.1)

betas <- coef(mnl.1)
varcov <- vcov(mnl.1)
# se <- sqrt(diag(mnl.1))

# betas[1]
# 
# # broom::tidy(mnl.1)
# stargazer(mnl.1, type = "text")

#Let’s set up the simulation.

set.seed(461982)
n <- 1000
sim.pred <- MASS::mvrnorm(n, betas, varcov)


#Extract the variable names.
c.n.2 <- names(betas)[names(betas) %>% str_detect("Traditionalist Populism")]
c.n.3 <- names(betas)[names(betas) %>% str_detect("Progressive Populism")]


#Set the independent variables to some meaningful values storing them into an object.

means <- c(mean(mlogit_ess$econ_unsat, na.rm = T), 
           mean(mlogit_ess$econ_insec, na.rm = T), 
           1, # Employed
           mean(mlogit_ess$anti_imm, na.rm = T),
           mean(mlogit_ess$authoritarianism, na.rm = T), 
           mean(mlogit_ess$globalism, na.rm = T), 
           mean(mlogit_ess$demsat, na.rm = T), 
           mean(mlogit_ess$trust_gov, na.rm = T),
           mean(mlogit_ess$govsat, na.rm = T), 
           mean(mlogit_ess$age, na.rm = T), 
           mean(mlogit_ess$educ, na.rm = T),
           1, #sex
           mean(mlogit_ess$lrscale, na.rm = T),
           1, #ethnic
           mean(mlogit_ess$religion, na.rm = T), 
           mean(mlogit_ess$rural, na.rm = T))

means <- data.frame(econ_unsat = mean(mlogit_ess$econ_unsat, na.rm = T), 
           econ_insec = mean(mlogit_ess$econ_insec, na.rm = T), 
           unemployed = 1, # Employed
#           anti_imm = mean(mlogit_ess$anti_imm, na.rm = T),
           authoritarianism = mean(mlogit_ess$authoritarianism, na.rm = T), 
           globalism = mean(mlogit_ess$globalism, na.rm = T), 
           demsat = mean(mlogit_ess$demsat, na.rm = T), 
           trust_gov = mean(mlogit_ess$trust_gov, na.rm = T),
           govsat = mean(mlogit_ess$govsat, na.rm = T), 
           age = mean(mlogit_ess$age, na.rm = T), 
           educ = mean(mlogit_ess$educ, na.rm = T),
           sex = 1,
           lrscale = mean(mlogit_ess$lrscale, na.rm = T),
           ethnic = 1,
           religion = mean(mlogit_ess$religion, na.rm = T), 
           rural = mean(mlogit_ess$rural, na.rm = T))
)
           
x.range <- seq(min(mlogit_ess$anti_imm, na.rm = T), max(mlogit_ess$anti_imm, na.rm = T), along.with = min(mlogit_ess$anti_imm, na.rm = T):max(mlogit_ess$anti_imm, na.rm = T))

#Create empty matrices for the predictions.
res.1 <- matrix(NA, nrow = length(x.range), ncol = 4)
res.2 <- matrix(NA, nrow = length(x.range), ncol = 4)
res.3 <- matrix(NA, nrow = length(x.range), ncol = 4)

#Run the loop.
i = 2
for(i in 1:length(x.range)){
  a <- x.range[i]
  # Generate vectors of predictions on the linear predictor
  b.2 <- sim.pred[, c.n.2]%*%c(1, means,a)
  b.3 <- sim.pred[, c.n.3]%*%c(1, means,a)
  # Generate vectors of probabilities (notice that category 1 is the baseline)
  p.1 <- 1/(1 + exp(b.2) + exp(b.3))
  p.2 <- exp(b.2)/(1 + exp(b.2) + exp(b.3))
  p.3 <- exp(b.3)/(1 + exp(b.2) + exp(b.3))
  # Put means and CIs in 4 matrices
  res.1[i, ] <- c(a,
                  mean(p.1), 
                  quantile(p.1, p = 0.025),
                  quantile(p.1, p = 0.975))
  res.2[i, ] <- c(a,
                  mean(p.2), 
                  quantile(p.2, p = 0.025),
                  quantile(p.2, p = 0.975))
  res.3[i, ] <- c(a,
                  mean(p.3), 
                  quantile(p.3, p = 0.025),
                  quantile(p.3, p = 0.975))
}

data.sim <- data.frame(rbind(res.1, res.2, res.3))
names(data.sim) <- c("X", "Y", "LO", "HI")
data.sim$GR <- rep(c("Establishment", "Traditional Populism", "Progressive Populism"), each = length(x.range))


install.packages("VGAM")
VGAM::margeff(test)


ggplot(data.sim, aes(x = X, y = Y, group = GR)) +
  geom_line() +
  geom_ribbon(aes(ymin = LO, ymax = HI), fill = "gray70", alpha = 0.5) +
  facet_wrap(~ GR, ncol = 3) +
  scale_y_continuous(limits = c(0,1)) +
#  scale_x_continuous(breaks = seq(0, 10, by = 1)) +
  xlab("Religiosity") +
  ylab("Probability to vote for party") +
  theme_bw()
```


# Trees

```{r}
tre_ess <- mlogit_ess %>% 
  mutate(progressive_populism = ifelse(target == "Progressive Populism", 1, 0)) %>% 
  mutate(trad_populism = ifelse(target == "Traditionalist Populism", 1, 0)) %>%    
  select(progressive_populism, unemployed, anti_imm)  
  
#pacman::p_load(FFTrees)
fft_populism <- FFTrees(formula = progressive_populism ~.,
                    data = tre_ess,
                    main = "Populism",
                    decision.labels = c("Establishment", "Traditional Populism"))

plot(fft_populism)



```

# Mlogit -TABLE

```{r}
library(nnet)
mlogit_ess$populism2 <- relevel(mlogit_ess$target, ref = "Establishment")

test <- multinom(
  populism2 ~
    # economic deprivation
    econ_insec + unemployed + welfare +
    # cultural backlash
    anti_imm + opendim + selfdim + globalism + govsat +
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + east + north + south +
    year_2016 + year_2014 + year_2012,
    #data
    data = mlogit_ess, weights = pweight)

# screenreg(test)
# stargazer::stargazer(test, type = "text")

multi1.rrr <- exp(coef(test))

stargazer::stargazer(test, type = "text",coef=list(multi1.rrr), p.auto=FALSE)

summary(test)

#car::Anova(test)

# PseudoR2(test)
# 
# regm2 <- step(test)
# # 
# # GGally::ggcoef(regm2, exponentiate = TRUE) + facet_grid(~y.level)
# 
# tidy(regm2, exponentiate = T, conf.int = TRUE) %>% 
#   # filter(term != "south") %>% 
#   # filter(term != "east") %>% 
#   # filter(term != "north") %>% 
#   # filter(term != "(Intercept)") %>% 
#   ggplot() + aes(x = term, y = estimate, 
#                  ymin = conf.low, 
#                  ymax = conf.high, color = y.level) + 
#   geom_hline(yintercept = 1, 
#              color = "gray25", 
#              linetype = "dotted") + 
#   geom_errorbar(position = position_dodge(0.5), width = 0) + 
#   geom_point(position = position_dodge(width = 0.5)) + scale_y_log10() + 
#   coord_flip()

```

## opendim-selfdim

```{r}

load("data/mlogit_ess_adds.Rdata")


library(nnet)
mlogit_ess_adds$populism2 <- relevel(mlogit_ess_adds$target, ref = "Establishment")

controls <- multinom(
  populism2 ~
    # economic deprivation
    #econ_unsat + econ_insec + unemployed + welfare +
    # cultural backlash
    #anti_imm + opendim + selfdim +
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east + north + south +
    year_2016 + year_2014 + year_2012,
    #data
    data = mlogit_ess_adds, weights = pweight)

# controls_tidy <- step(controls) %>% 
#   tidy(exponentiate = T, conf.int = TRUE) %>% 
#   mutate(model = "control") %>% 
#   mutate(stars = get_p_values(p.value)) %>% 
#   mutate(mcfadden = round(PseudoR2(controls), 2)) 

  # cowplot::add_sub(gg_coefs, 
  #                  label = paste(expression(paste("McFadden's Pseudo R^2:")), 
  #                                           gg_coefs$mcfadden), 
  #                  x = 0, hjust = 0) %>%
  # cowplot::ggdraw()

fit1 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south + year_2016 + year_2014 + year_2012 +
    # economic deprivation
    econ_insec + unemployed + welfare,
    # cultural backlash
    #anti_imm + opendim + selfdim + globalism + govsat +
    #data
    data = mlogit_ess_adds, weights = pweight)

# fit1_tidy <- step(fit1) %>% 
#   tidy(exponentiate = T, conf.int = TRUE) %>% 
#   mutate(model = "control") %>% 
#   mutate(stars = get_p_values(p.value)) %>% 
#   mutate(mcfadden = round(PseudoR2(fit1), 2)) 


fit2 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south +  year_2016 + year_2014 + year_2012+
    # economic deprivation
    #econ_unsat + econ_insec + unemployed + welfare +
    # cultural backlash
    anti_imm + opendim + selfdim,
    #data
    data = mlogit_ess_adds, weights = pweight)


fit3 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south +   year_2016 + year_2014 + year_2012+
    # economic deprivation
    econ_insec + unemployed + welfare +
    # cultural backlash
    anti_imm + opendim + selfdim,
    #data
    data = mlogit_ess_adds, weights = pweight)

fit4 <- multinom(
  populism2 ~
    # control variables
    age + educ + sex + lrscale + ethnic + religion + rural + globalism + govsat + east +
    north + south +   year_2016 + year_2014 + year_2012+
    # economic deprivation
    econ_insec + unemployed + welfare +
    # cultural backlash
    anti_imm + opendim + selfdim +
    # interactions
    econ_insec * anti_imm + 
    #econ_insec * conservation + 
    unemployed * opendim,
    #data
    data = mlogit_ess_adds, weights = pweight)

# PseudoR2(controls)
# PseudoR2(fit1)
# PseudoR2(fit2)
# PseudoR2(fit3)
# PseudoR2(fit4)
#

# tidy(fit4) %>%
#   filter(term %nin% c("(Intercept)", "age", "educ", "sex",
#                      "lrscale", "ethnic", "religion", "rural",
#                      "globalism", "govsat", "east", "north", "south",
#                      "year_2016", "year_2014", "year_2012", "econ_insec",
#                      "unemployed", "welfare", "anti_imm", "openness",
#                      "conservation", "selfenhance", "selftrans")) %>%
#   mutate(stars = get_p_values(p.value))

# econ_insec:anti_imm only traditionalist populism
# econ_insec:conservation only traditionalist populism - weak
# unemployed:selftrans only progressive populism - weak

```


### making the adds plot

```{r}



control_variables <- c("age", "educ", "sex", "lrscale", "ethnic", 
  "religion", "rural", "south", "east", "north", 
             "globalism", "govsat")
  
economic <- c("econ_insec", "unemployed", "welfare")

cultural <- c("anti_imm", "opendim", 
             "selfdim")

all_mods %>% 
  filter(term != "south") %>% 
  filter(term != "east") %>% 
  filter(term != "north") %>% 
  filter(term != "year_2016") %>% 
  filter(term != "year_2014") %>% 
  filter(term != "year_2012") %>% 
  filter(term != "(Intercept)") %>% 
  mutate(variable_groups = case_when(
    term %in% control_variables ~ "Controls",
    term %in% economic ~ "Economic",
    term %in% cultural ~ "Cultural",
    TRUE ~ "Interactions"
  )) %>% 
  mutate(model = case_when(
    model == "control" ~ "Model 1 (Controls)",
    model == "model1" ~ "Model 2 (Economic)",
    model == "model2" ~ "Model 3 (Culture)",
    model == "model3" ~ "Model 4 (Economic + Culture)",
    model == "model4" ~ "Model 5 (Interactions)"
  )) %>%
  mutate(term = case_when(
    term == "anti_imm" ~ "Anti-Immigration Sentiment", 
    term == "opendim" ~ "Openness", 
    term == "selfdim" ~ "Selfishness", 
    # term == "selfenhance" ~ "Self-Enhancement", 
    # term == "selftrans" ~ "Self-Transcendence", 
    term == "econ_insec" ~ "Economic Insecurity", 
    term == "unemployed" ~ "Unemployed (0/1)", 
    term == "welfare" ~ "Welfare (0/1)",
    term == "age" ~ "Age", 
    term == "educ" ~ "Education", 
    term == "sex" ~ "Female (0/1)", 
    term == "lrscale" ~ "Left-Right Scale", 
    term == "ethnic" ~ "Ethnic Minority (0/1)", 
    term == "religion" ~ "Religiosity", 
    term == "rural" ~ "Rural (0/1)",  
    term == "globalism" ~ "Trust in Global Governance", 
    term == "govsat" ~ "Government Satisfaction",
    term == "econ_insec:anti_imm" ~ "Econ. Insec. X Anti-Immigration",
#    term == "econ_insec:conservation" ~ "Economic Insecurity X Conservation",
    term == "unemployed:opendim" ~ "Unemployed X Openness"
  )) %>% 
  mutate(term = fct_relevel(term, levels = c("Unemployed X Openness",
                              #"Economic Insecurity X Conservation",
                              "Econ. Insec. X Anti-Immigration",
                              "Anti-Immigration Sentiment", "Openness", 
                              "Selfishness", 
                              "Econ Insecurity", "Unemployed (0/1)", 
                              "Welfare (0/1)",
                              "Age", "Education", "Female (0/1)", "Left-Right Scale", 
                              "Ethnic Minority (0/1)", "Religiosity", "Rural (0/1)",
                              "Trust in Global Governance", "Government Satisfaction"))) %>% 
  mutate(`Variable Groups` = variable_groups) %>% 
  mutate(`Y-Level` = y.level) %>% 
  ggplot() + aes(x = term, y = estimate, 
                 ymin = conf.low, 
                 ymax = conf.high, 
                 color = `Y-Level`) + 
  geom_hline(yintercept = 1, 
             color = "gray25", 
             linetype = "dotted") + 
  geom_errorbar(position = position_dodge(0.5), width = 0) + 
  geom_point(aes(shape = `Variable Groups`), 
             position = position_dodge(width = 0.5)) + 
  scale_y_log10() +
  geom_vline(xintercept = seq(1.5, length(unique(all_mods$term)) - 0.5, 1), 
             lwd = 1, colour = "grey", alpha = 0.7) +
  geom_text(aes(x = term, 
                y = estimate + 0.01,
                label = paste(sprintf('%.2f', estimate, 2), stars)), 
                position = position_dodge(1.2), show.legend = F) +
  coord_flip() +
  guides(point = F) +
  scale_y_continuous(breaks = seq(-2, 2, by = 0.3), limits = c(.5, 1.9)) +
  ggthemes::theme_hc() +
  ggthemes::scale_fill_fivethirtyeight() +
  cowplot::background_grid(major = 'y', minor = "none") + # add thin horizontal lines
  cowplot::panel_border() +# and a border around each panel
  xlab("") +
  ylab("Odds Ratios") + 
  # labs(caption = paste(expression("McFadden's Pseudo R^2:")), 
  #                                          gg_coefs$mcfadden) +
  facet_grid(~ model, scales = "fixed") +
  annotate("label", x = -1, y = 1.25, label = paste("McFadden's R2.:",
                                                  mcfaddies,
                                                  "\n N:", rep(nrow(mlogit_ess), 5),
                                                  "\n LRT:", anovies)) +
  # annotate("label", x = 7.5, y = 0.7, label = "Controls") +
  # annotate("label", x = 11.5, y = 0.7, label = "Economic") +
  # annotate("label", x = 18.5, y = 0.7, label = "Cultural") +
  # annotate("label", x = 19.5, y = 0.7, label = "Interactions") +
  # expand_limits(x = 21) + 
  geom_text(aes(x = -2, label = ""), show.legend = F) +
  scale_x_discrete(limits = c("Unemployed X Openness",
                              #"Economic Insecurity X Conservation",
                              "Econ. Insec. X Anti-Immigration",
                              "Interactions",
                              "Anti-Immigration Sentiment", "Openness", 
                              "Selfishness",
                              "Culture",
                              "Economic Insecurity", "Unemployed (0/1)", 
                              "Welfare (0/1)",
                              "Economics",
                              "Age", "Education", "Female (0/1)", "Left-Right Scale", 
                              "Ethnic Minority (0/1)", "Religiosity", "Rural (0/1)",
                              "Trust in Global Governance", "Government Satisfaction",
                              "Controls"), 
                   labels = c("Controls"=expression(displaystyle(bold(Controls))), 
                            "Economics"=expression(displaystyle(bold(Economics))),
                            "Culture"=expression(displaystyle(bold(Culture))), 
                            "Interactions"=expression(displaystyle(bold(Interactions))), 
                            parse=TRUE))

ggsave(filename = "images/onebigmotherfucker2.png", width = 15, height = 10)
 
# multi1.rrr <- exp(coef(test))
# 
# stargazer::stargazer(test, type="text",coef=list(multi1.rrr), p.auto=FALSE)

# TODO Make a map (or two)
# TODO Make stuff citeable
# TODO Decide which interactions
# TODO Make probability plots

unique(all_mods$term)
```




#### anti-immigration

```{r}
# test <- multinom(
#   populism2 ~
#     # economic deprivation
#     econ_unsat + econ_insec + unemployed +
#     # cultural backlash
#     anti_imm + opendim + selfdim + globalism + govsat +
#     # control variables
#     age + educ + sex + lrscale + ethnic + religion + rural + east + north + south,
#     #data
#     data = mlogit_ess, weights = pweight)
# 
# 
# screenreg(test)

## Automatic Picking of Scales
#from_to <- as.vector(names(table(ess_mlogit$openness)))
#fit.eff <- effects::effect("openness", 
#                           test, 
#                           xlevels = list(openness = from_to))
## or
#fit.eff <- effects::effect("openness", 
#                           test)
## or put them in by hand
fit.eff <- effects::effect("anti_imm", 
                           fit3, 
                           xlevels = list(anti_imm = c(1:10)))

df <- fit.eff$prob %>% 
  as.data.frame()

data.frame(diff(as.matrix(df))) %>% 
  #janitor::clean_names() %>% 
  summarise_all(mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) -> ame_dat

anti_imm_prob <- data.frame(fit.eff$model.matrix, 
           fit.eff$prob, 
           fit.eff$lower.prob, 
           fit.eff$upper.prob) %>% 
  select(anti_imm, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, -anti_imm, 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>%
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
  ggplot(aes(anti_imm, value, colour = type)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    geom_text(data = ame_dat, aes(x = 2, y = -Inf, label = V1),
              hjust   = -0.1,
              vjust   = -1 
              #nudge_y = 15, 
              #nudge_x = 15
              ) +
    geom_jitter(alpha = 0.01) +
  labs(y = "Marginal Probability") +
  guides(color = F) +
    ggthemes::theme_hc() +
    ggthemes::scale_color_fivethirtyeight() + 
    facet_wrap(~type, scales = "free", nrow = 1) +
    scale_y_continuous(breaks = seq(0,100, by = 1), 
                     labels = paste0(seq(0,100, by = 1),"%")) +
    scale_x_continuous(breaks = seq(0,10, by = 1), 
                     labels = seq(0,10, by = 1)) +
  xlab("Anti-Immigration Sentiment") 

ggsave(anti_imm_prob, filename = "images/anti_imm_prob.png", width = 12, height = 4)
```

#### Openness

```{r}
fit.eff <- effects::effect("openness", 
                           fit3, 
                           xlevels = list(openness = c(1:10)))

df <- fit.eff$prob %>% 
  as.data.frame()

data.frame(diff(as.matrix(df))) %>% 
  #janitor::clean_names() %>% 
  summarise_all(mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) -> ame_dat

openness_prob <- data.frame(fit.eff$model.matrix, 
           fit.eff$prob, 
           fit.eff$lower.prob, 
           fit.eff$upper.prob) %>% 
  select(openness, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, -openness, 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>%
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
  ggplot(aes(openness, value, colour = type)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    geom_text(data = ame_dat, aes(x = 2, y = -Inf, label = V1),
              hjust   = -0.1,
              vjust   = -1 
              #nudge_y = 15, 
              #nudge_x = 15
              ) +
    geom_jitter(alpha = 0.01) +
  labs(y = "") +
  guides(color = F) +
    ggthemes::theme_hc() +
    ggthemes::scale_color_fivethirtyeight() + 
    facet_wrap(~type, scales = "free", nrow = 1) +
    scale_y_continuous(breaks = seq(0,100, by = 1), 
                     labels = paste0(seq(0,100, by = 1),"%")) +
    scale_x_continuous(breaks = seq(0,10, by = 1), 
                     labels = seq(0,10, by = 1)) +
  xlab("Openness Scale") 

ggsave(openness_prob, filename = "images/openness.png", width = 12, height = 4)
```

#### Conservatism

```{r}
fit.eff <- effects::effect("conservation", 
                           fit3, 
                           xlevels = list(conservation = c(1:10)))

df <- fit.eff$prob %>% 
  as.data.frame()

data.frame(diff(as.matrix(df))) %>% 
  #janitor::clean_names() %>% 
  summarise_all(mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) -> ame_dat

conservation_prob <- data.frame(fit.eff$model.matrix, 
           fit.eff$prob, 
           fit.eff$lower.prob, 
           fit.eff$upper.prob) %>% 
  select(conservation, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, -conservation, 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>%
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
  ggplot(aes(conservation, value, colour = type)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    geom_text(data = ame_dat, aes(x = 2, y = -Inf, label = V1),
              hjust   = -0.1,
              vjust   = -1 
              #nudge_y = 15, 
              #nudge_x = 15
              ) +
    geom_jitter(alpha = 0.01) +
  labs(y = "") +
  guides(color = F) +
    ggthemes::theme_hc() +
    ggthemes::scale_color_fivethirtyeight() + 
    facet_wrap(~type, scales = "free", nrow = 1) +
    scale_y_continuous(breaks = seq(0,100, by = 1), 
                     labels = paste0(seq(0,100, by = 1),"%")) +
    scale_x_continuous(breaks = seq(0,10, by = 1), 
                     labels = seq(0,10, by = 1)) +
  xlab("Conservation Scale") 

ggsave(conservation_prob, filename = "images/conservation.png", width = 12, height = 4)
```


#### selfenhancment

```{r}
fit.eff <- effects::effect("selfenhance", 
                           fit3, 
                           xlevels = list(selfenhance = c(1:10)))

df <- fit.eff$prob %>% 
  as.data.frame()

data.frame(diff(as.matrix(df))) %>% 
  #janitor::clean_names() %>% 
  summarise_all(mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) -> ame_dat

selfenhance_prob <- data.frame(fit.eff$model.matrix, 
           fit.eff$prob, 
           fit.eff$lower.prob, 
           fit.eff$upper.prob) %>% 
  select(selfenhance, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, -selfenhance, 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>%
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
  ggplot(aes(selfenhance, value, colour = type)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    geom_text(data = ame_dat, aes(x = 2, y = -Inf, label = V1),
              hjust   = -0.1,
              vjust   = -1 
              #nudge_y = 15, 
              #nudge_x = 15
              ) +
    geom_jitter(alpha = 0.01) +
  labs(y = "") +
  guides(color = F) +
    ggthemes::theme_hc() +
    ggthemes::scale_color_fivethirtyeight() + 
    facet_wrap(~type, scales = "free", nrow = 1) +
    scale_y_continuous(breaks = seq(0,100, by = 1), 
                     labels = paste0(seq(0,100, by = 1),"%")) +
    scale_x_continuous(breaks = seq(0,10, by = 1), 
                     labels = seq(0,10, by = 1)) +
  xlab("Self-Transcendence Scale") 

ggsave(selfenhance_prob, filename = "images/selfenhance_prob.png", width = 12, height = 4)
```


#### selftrans

```{r}
fit.eff <- effects::effect("selftrans", 
                           fit3, 
                           xlevels = list(selftrans = c(1:10)))

df <- fit.eff$prob %>% 
  as.data.frame()

data.frame(diff(as.matrix(df))) %>% 
  #janitor::clean_names() %>% 
  summarise_all(mean) %>% 
  t() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "type") %>% 
  mutate(type = case_when(
    str_detect(type, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(type, "Progressive") ~ "Progressive Populism",
    str_detect(type, "Establishment") ~ "Establishment"
  )) %>% 
  mutate(V1 = paste0("AME = ", sprintf('%.2f', V1*100, 2),"%")) -> ame_dat

selftrans_prob <- data.frame(fit.eff$model.matrix, 
           fit.eff$prob, 
           fit.eff$lower.prob, 
           fit.eff$upper.prob) %>% 
  select(selftrans, prob.Traditionalist.Populism, prob.Establishment,
         prob.Progressive.Populism, L.prob.Establishment, 
         U.prob.Establishment, L.prob.Traditionalist.Populism, 
         U.prob.Traditionalist.Populism, L.prob.Progressive.Populism, 
         U.prob.Progressive.Populism, -X.Intercept.) %>% 
  gather(key = variable, value = value, -selftrans, 
           -L.prob.Establishment, -U.prob.Establishment,
           -L.prob.Traditionalist.Populism, -U.prob.Traditionalist.Populism,
         -L.prob.Progressive.Populism, -U.prob.Progressive.Populism) %>% 
  mutate(type = case_when(
    str_detect(variable, "Traditionalist") ~ "Traditionalist Populism",
    str_detect(variable, "Progressive") ~ "Progressive Populism",
    str_detect(variable, "Establishment") ~ "Establishment"
  )) %>%
    mutate(lower = case_when(
    type == "Establishment" ~ L.prob.Establishment*100,
    type == "Progressive Populism" ~ L.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ L.prob.Traditionalist.Populism*100
  )) %>% 
    mutate(upper = case_when(
    type == "Establishment" ~ U.prob.Establishment*100,
    type == "Progressive Populism" ~ U.prob.Progressive.Populism*100,
    type == "Traditionalist Populism" ~ U.prob.Traditionalist.Populism*100
  )) %>% 
  mutate(value = value * 100) %>% 
  ggplot(aes(selftrans, value, colour = type)) +
    geom_smooth(method = "loess", 
               #formula = y ~ s(x, k = 1), 
               size = 1.2, se = F) +
    geom_ribbon(aes(ymin = lower, ymax = upper), 
               fill = "gray70", alpha = 0.5, colour=NA) +
    geom_text(data = ame_dat, aes(x = 2, y = -Inf, label = V1),
              hjust   = -0.1,
              vjust   = -1 
              #nudge_y = 15, 
              #nudge_x = 15
              ) +
    geom_jitter(alpha = 0.01) +
  labs(y = "") +
  guides(color = F) +
    ggthemes::theme_hc() +
    ggthemes::scale_color_fivethirtyeight() + 
    facet_wrap(~type, scales = "free", nrow = 1) +
    scale_y_continuous(breaks = seq(0,100, by = 1), 
                     labels = paste0(seq(0,100, by = 1),"%")) +
    scale_x_continuous(breaks = seq(0,10, by = 1), 
                     labels = seq(0,10, by = 1)) +
  xlab("Self-Transcendence Scale") 

ggsave(selftrans_prob, filename = "images/selftrans_prob.png", width = 12, height = 4)
```

##### Cowplotting it together

```{r}
cultural_probs <- cowplot::plot_grid(anti_imm_prob, openness_prob, 
                               conservation_prob, selfenhance_prob, 
                               selftrans_prob,
                      labels = c("A", "B", "C", "D", "E"), ncol = 1)

cowplot::save_plot("images/cultural_probs.png", cultural_probs,
          #ncol = 2, # we're saving a grid plot of 2 columns
          ncol = 2, # and 2 rows
          base_height = 10,
          base_width = 6,
          # each individual subplot should have an aspect ratio of 1.3
          base_aspect_ratio = 1.3
          )
```
